% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_unit_options}
\alias{area_unit_options}
\title{Area units (vector)}
\format{
A character vector with 35 names, plural names, and aliases for area
units.
}
\usage{
area_unit_options
}
\description{
A vector of supported area units derived from \link{dist_units} and
\link[units:valid_udunits]{units::valid_udunits}.
}
\keyword{datasets}
