% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_sf_coverage.R
\name{bind_sf_coverage}
\alias{bind_sf_coverage}
\alias{st_make_valid_coverage}
\alias{st_make_valid_union}
\title{Row bind coverage for a feature based on a coverage feature}
\usage{
bind_sf_coverage(
  x,
  coverage,
  coverage_nm = NULL,
  .id = "name",
  x_arg = caller_arg(x),
  coverage_arg = caller_arg(coverage),
  id_arg = caller_arg(.id),
  combine = FALSE,
  error_call = caller_env()
)

st_make_valid_coverage(
  x,
  y,
  combine = FALSE,
  is_coverage = TRUE,
  .fn = sf::st_difference
)

st_make_valid_union(x, combine = FALSE, is_coverage = TRUE)
}
\arguments{
\item{x}{A input sf object}

\item{coverage}{A sf, sfc, or bbox object that covers x.}

\item{coverage_nm}{Name to use for coverage feature assigned as value for .id
column, Default: \code{NULL}}

\item{.id}{.id column name, must be present in x Default: 'name'}

\item{x_arg, coverage_arg, id_arg}{Argument names for error messages if
\code{\link[=bind_sf_coverage]{bind_sf_coverage()}} is used as a wrapper function.}

\item{combine}{If \code{TRUE}, use \code{\link[sf:st_combine]{sf::st_combine()}} before passing x to
\code{\link[sf:st_union]{sf::st_union()}}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{y}{Second parameter for \code{\link[=st_make_valid_coverage]{st_make_valid_coverage()}}}

\item{is_coverage}{logical; if \code{TRUE}, use an optimized algorithm for features that form a polygonal coverage (have no overlaps)}

\item{.fn}{Defaults to sf::st_difference}
}
\value{
A sf object or a sfc object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{\link[=bind_sf_coverage]{bind_sf_coverage()}} and the related helper functions
\code{\link[=st_make_valid_coverage]{st_make_valid_coverage()}} and \code{\link[=st_make_valid_union]{st_make_valid_union()}} take a sf object and
create a non-intersecting coverage area of a surrounding geography defined by
a second sf, sfc, or bbox object. This was originally developed for the
\href{https://elipousson.github.io/getACS/reference/make_area_xwalk.html}{make_area_xwalk()}
function in the getACS package.
}
\seealso{
\code{\link[vctrs:vec_bind]{vctrs::vec_bind()}}
\code{\link[sf:valid]{sf::valid()}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
\keyword{internal}
