% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_format.R
\name{cli_format.sf}
\alias{cli_format.sf}
\title{cli_format style for sf objects}
\usage{
\method{cli_format}{sf}(x, style = NULL, ...)
}
\arguments{
\item{x}{A sf object to format with the cli_format.sf method.}

\item{style}{Not in use for the 'sf' class method.}

\item{...}{Not in use for the 'sf' class method.}
}
\description{
cli_format style for sf objects
}
\examples{
\dontrun{
if (interactive()) {
  library(cli)
  library(sf)
  cliExtras::register_cli_format("sf", cli_format.sf)
  nc <- st_read(system.file("shape/nc.shp", package = "sf"))
  cli_text("`nc` is a {.val {nc}}")
}
}
}
\seealso{
\code{\link[cliExtras:register_cli_format]{cliExtras::register_cli_format()}}
}
