% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dist_units.R
\name{convert_dist_units}
\alias{convert_dist_units}
\title{Convert distance (and area) values between different units}
\usage{
convert_dist_units(
  dist,
  from = NULL,
  to = "meter",
  drop = FALSE,
  digits = NULL
)
}
\arguments{
\item{dist}{Numeric or units object}

\item{from}{Existing unit for dist, Default: \code{NULL}. If dist is a units
object, the numerator is used as "from"}

\item{to}{Unit to convert distance to, Default: 'meter'}

\item{drop}{If \code{TRUE}, return numeric. If \code{FALSE}, return class units object.}

\item{digits}{Number of digits to include in result; defaults to \code{NULL}.}
}
\value{
Object created by \code{\link[units:set_units]{units::set_units()}}
}
\description{
Convert distance (and area) values between different units
}
\seealso{
Other dist:
\code{\link[=compare_dist]{compare_dist()}},
\code{\link[=convert_dist_scale]{convert_dist_scale()}},
\code{\link{get_measurements}},
\code{\link[=is_dist_units]{is_dist_units()}},
\code{\link[=sf_bbox_dist]{sf_bbox_dist()}}
}
\concept{dist}
