% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{count_features}
\alias{count_features}
\title{Count simple features based on relationship with a second simple feature object}
\usage{
count_features(
  x = NULL,
  y = NULL,
  nm = "data",
  join = NULL,
  .id = "name",
  by = NULL,
  count = NULL,
  sort = FALSE,
  name = NULL,
  geometry = "y",
  ...
)
}
\arguments{
\item{x}{Data frame or \code{sf} object, Default: \code{NULL}}

\item{y}{Length 1 named \code{sf} list (name of y is used as count if count is
\code{NULL}) or a \code{sf} object if nm is not \code{NULL}. y must include a column name
matching the value of .id. If y is \code{NULL}, count is required. Default: \code{NULL}}

\item{nm}{Vector of names to use with \code{\link[=as_sf_list]{as_sf_list()}} to convert y to an sf list
(if y is not already an sf list). Defaults to "data".}

\item{join}{geometry predicate function; defaults to \code{NULL}, set to
\code{\link[sf:st_intersects]{sf::st_intersects()}} if y contains only POLYGON or MULTIPOLYGON
objects or \code{\link[sf:st_nearest_feature]{sf::st_nearest_feature()}} if y contains other types.}

\item{.id}{Column name with the values that should be added as a column to
the input \code{sf} object.}

\item{by}{A character vector of variables to join by passed to
\code{\link[dplyr:left_join]{dplyr::left_join()}}.}

\item{count}{Name of column to count. If \code{NULL}, count is set to \code{names(y)}
(assuming y is a length 1 named \code{sf} list).}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.

If omitted, it will default to \code{n}. If there's already a column called \code{n},
it will use \code{nn}. If there's a column called \code{n} and \code{nn}, it'll use
\code{nnn}, and so on, adding \code{n}s until it gets a new name.}

\item{geometry}{If "y", replace x geometry with y geometry joining based on
by. If by is \code{NULL}, by is set to the same value as count.}

\item{...}{Additional parameters passed to \code{\link[=st_join_ext]{st_join_ext()}}}
}
\description{
Use \code{\link[=st_join_ext]{st_join_ext()}} and \code{\link[dplyr:count]{dplyr::count()}} to count features in x based on
their spatial relationship with y. Very similar to \code{\link[=count_sf_ext]{count_sf_ext()}} so they
may be merged in the future.
}
