% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_unit_options}
\alias{dist_unit_options}
\title{Distance units (vector)}
\format{
A character vector with 86 names, plural names, aliases, and symbols
for distance units.
}
\usage{
dist_unit_options
}
\description{
A vector of supported distance units pulled from \code{dist_units}.
}
\keyword{datasets}
