% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_units}
\alias{dist_units}
\title{Distance units (data frame)}
\format{
A data frame with 33 rows and 12 variables:
\describe{
\item{\code{symbol}}{symbols}
\item{\code{symbol_aliases}}{symbol aliases}
\item{\code{name_singular}}{singular names}
\item{\code{name_singular_aliases}}{singular name aliases}
\item{\code{name_plural}}{character plural names}
\item{\code{name_plural_aliases}}{plural name aliases}
\item{\code{def}}{short definition}
\item{\code{definition}}{definition}
\item{\code{comment}}{comment}
\item{\code{dimensionless}}{logical indicator for dimensionless units}
\item{\code{source_xml}}{source XML}
\item{\code{unit_opts}}{character vector with symbols, singular, and plural
names for the unit}
}
}
\usage{
dist_units
}
\description{
A subset of units supported by the units package accessible through the
\code{\link[units:valid_udunits]{units::valid_udunits()}} function.
}
\keyword{datasets}
