% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords}
\alias{get_coords}
\alias{st_coords}
\alias{get_minmax}
\alias{st_coords_minmax}
\title{Get coordinates for a simple feature or bounding box object}
\usage{
get_coords(
  x,
  coords = NULL,
  geometry = NULL,
  crs = NULL,
  keep_all = TRUE,
  drop = TRUE
)

get_minmax(x, crs = NULL, keep_all = TRUE, drop = TRUE)
}
\arguments{
\item{x}{`sf`, `bbox`, or `sfc` object}

\item{coords}{Column names to use for coordinates in results, Default: `NULL`;
which is set to c("lon", "lat") by [check_coords]}

\item{geometry}{geometry to use for coordinates "centroid", "surface point",
or alternatively "wkt"; defaults to `NULL` ("centroid")}

\item{crs}{Coordinate reference system to use for coordinates; defaults to `NULL`.}

\item{keep_all}{If `TRUE`, bind the coordinates columns to the provided object x,
Default: `TRUE`.}

\item{drop}{If `TRUE` and x is an sf object, drop the geometry Default: `TRUE`.}
}
\description{
An extended version of [sf::st_coordinates] that supports binding coordinates
to the object, optionally dropping the geometry, and returning wkt or a point
on surface (geometry = "surface point") instead of the centroid.
}
\details{
[get_minmax] get a bounding box for each feature (or group of features)
appends the xmin, ymin, xmax, and ymax values for each feature to the simple
feature object.
}
