% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{get_paper}
\alias{get_paper}
\title{Get standard paper and image sizes}
\usage{
get_paper(
  paper = "letter",
  orientation = "portrait",
  standard = NULL,
  series = NULL,
  size = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  ncol = 1,
  nrow = 1,
  gutter = 0,
  bbox = NULL,
  margin = NULL,
  ...
)
}
\arguments{
\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}

\item{standard}{Size standard, "ANSI", "ISO", "British Imperial", "JIS",
"USPS", "Facebook", "Instagram", or "Twitter".}

\item{series}{Size series (e.g. A), Default: \code{NULL}}

\item{size}{Size number (only used for "ISO" and "JIS" series). Standard,
series, and size may all be required to return a single paper when using
these parameters.}

\item{width, height}{Width and height in units, Default: \code{NULL}.}

\item{units}{Paper size units, either "in", "mm", or "px"; defaults to \code{NULL}
(using "in" if width or height are provided).}

\item{ncol, nrow}{Number of expected columns and rows in paper; used to
determine row_height and section_asp in paper data frame returned by
get_paper if nrow or ncol is greater than 1; defaults to \code{NULL}.}

\item{gutter}{Gutter distance in units. Gutter is used as the spacing between
nrow and columns (variable spacing is not currently supported); defaults to 0.}

\item{bbox}{A bounding box to use to get orientation using \code{\link[=sf_bbox_asp]{sf_bbox_asp()}} with orientation = TRUE.}

\item{margin}{A numeric vector or ggplot2 margin object.}

\item{...}{Additional parameters passed to get_margin. plot_width can only be
passed in these parameters if paper has only a single row. margin is returned as a list column.}
}
\value{
Data frame with one or more paper/image sizes.
}
\description{
Use the "paper" parameter (matching name from \link{paper_sizes}), standard
(optionally including series and size) parameter, or width, height and units.
May return multiple paper sizes depending on parameters.
}
\details{
If margin is provided, a block_width, block_height, and block_asp are
calculated and included as columns in the returned data frame.

Paper can also be a data frame with "width", "height", "orientation", and
"units" columns.
}
\examples{
get_paper("letter")

get_paper(paper = NULL, standard = "ISO", series = "A", size = 4)

get_paper(width = 11, height = 17)
}
