% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scale.R
\name{get_scale}
\alias{get_scale}
\alias{get_standard_scale}
\title{Get standard scales and convert to scale distances}
\usage{
get_scale(scale = NULL, standard = NULL, series = NULL)
}
\arguments{
\item{scale}{Scale name}

\item{standard}{USGS, Engineering, or Architectural}

\item{series}{Map series}
}
\description{
Get standard scales and convert to scale distances
}
