% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dist_units.R
\name{is_dist_units}
\alias{is_dist_units}
\alias{diff_dist}
\alias{is_diff_dist}
\alias{is_same_dist}
\alias{is_longer}
\alias{is_shorter}
\alias{get_dist_units}
\alias{as_dist_units}
\alias{is_diff_area}
\alias{is_same_area}
\alias{is_same_units}
\title{General utility functions for working with distance units objects}
\usage{
is_dist_units(x)

is_diff_dist(x, y, units = NULL)

is_same_dist(x, y, dist = NULL, diff = FALSE, call = caller_env(), ...)

is_longer(x, y)

is_shorter(x, y)

get_dist_units(x, allow_null = TRUE, multiple = TRUE, quiet = FALSE)

as_dist_units(x, units = NULL, allow_null = FALSE, call = caller_env())

is_diff_area(x, y, units = NULL, combine = TRUE)

is_same_area(x, y, units = NULL, combine = TRUE, diff = FALSE, ...)

is_same_units(x, y = NULL)
}
\arguments{
\item{x, y}{objects to check}

\item{units}{For \link{is_diff_dist}, if x and y are both not units objects, use
units; default to \code{NULL}.}

\item{dist}{type of distance to compare if x and y are \code{sf}, \code{sfc}, or \code{bbox}
objects; "diagdist", "xdist", "ydist". defaults to \code{NULL}.}

\item{diff}{If \code{TRUE}, return results from \link{is_diff_dist} or \link{is_diff_area};
if \code{FALSE}, return logical indicator; defaults to \code{FALSE}}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{...}{Additional parameters passed to \code{\link[=all.equal]{all.equal()}}}

\item{allow_null}{If allow_null is \code{TRUE}, allow x to return a \code{NULL} value;
if \code{FALSE}, error on \code{NULL} values.}

\item{multiple}{If \code{TRUE} and x is a character vector with distance/area
units, \link{get_dist_units} may return multiple units. Passed to \link[rlang:arg_match]{rlang::arg_match}.}

\item{quiet}{If \code{TRUE}, suppress warning messages.}

\item{combine}{If \code{TRUE}, combine objects with \code{\link[sf:st_combine]{sf::st_combine()}} before
comparing area with \code{\link[=is_diff_area]{is_diff_area()}} or \code{\link[=is_same_area]{is_same_area()}}, defaults to
\code{TRUE}.}
}
\description{
\itemize{
\item \code{\link[=is_dist_units]{is_dist_units()}}: Is x a distance unit object?
\item \code{\link[=is_diff_dist]{is_diff_dist()}}: What is the difference between x and y distance?
\item \code{\link[=is_same_dist]{is_same_dist()}}: Is x the same distance as y? or does the bbox of x and bbox
of y have the same x, y, or diagonal distance?
\item \code{\link[=is_shorter]{is_shorter()}}, \code{\link[=is_longer]{is_longer()}}: Is x shorter or longer than y?
\item \code{\link[=is_same_area]{is_same_area()}}: do x and y have the same area?
\item \code{\link[=is_same_units]{is_same_units()}}: are x and y character strings that represent the same
units or objects that use the same units?
}
}
\details{
There are two additional functions that support these utility functions:
\itemize{
\item \code{\link[=get_dist_units]{get_dist_units()}}: Get the distance units from x (if x is a sf or units
objects or a character string from \link{dist_unit_options})
\item \code{\link[=as_dist_units]{as_dist_units()}}: Convert x to units using \link[units:as_units]{units::as_units}
}
}
\seealso{
Other dist:
\code{\link[=compare_dist]{compare_dist()}},
\code{\link[=convert_dist_scale]{convert_dist_scale()}},
\code{\link[=convert_dist_units]{convert_dist_units()}},
\code{\link{get_measurements}},
\code{\link[=sf_bbox_dist]{sf_bbox_dist()}}
}
\concept{dist}
