% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{is_sf}
\alias{is_sf}
\alias{is_sfg}
\alias{is_sfc}
\alias{is_bbox}
\alias{is_sf_list}
\alias{is_raster}
\alias{is_sp}
\alias{is_coords}
\alias{is_geo_coords}
\alias{is_wgs84}
\alias{is_same_crs}
\title{What is the class or spatial attributes of this feature?}
\usage{
is_sf(x, ext = FALSE, allow_null = FALSE, allow_list = FALSE)

is_sfg(x, allow_null = FALSE)

is_sfc(x, allow_null = FALSE)

is_bbox(x, allow_null = FALSE)

is_sf_list(x, named = FALSE, ext = FALSE, allow_null = FALSE)

is_raster(x, allow_null = FALSE)

is_sp(x, allow_null = FALSE)

is_geo_coords(x, allow_null = FALSE)

is_wgs84(x)

is_same_crs(x, y)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{ext}{If \code{TRUE}, check if x is a \code{sf}, \code{sfc}, or \code{bbox} class object or
not; defaults to \code{FALSE}. (used by \link{is_sf})}

\item{allow_null}{If \code{TRUE} and x is \code{NULL}, return \code{TRUE}; defaults to \code{FALSE}.}

\item{allow_list}{If \code{TRUE}, \link{is_sf} will return TRUE if x is a list of sf objects.}

\item{named}{If \code{TRUE}, check if sf list is named; defaults \code{FALSE}.}

\item{y}{An sf object or a character or numeric object supported by
\link[sf:st_crs]{sf::st_crs} that can be compared to x. (used by \link{is_same_crs})}
}
\description{
What is the class or spatial attributes of this feature?
}
\details{
\itemize{
\item \link{is_sf}: is x a \code{sf} class object?
\item \link{is_sfc}: is x is a \code{sfc} class object?
\item \link{is_bbox}: is x is a \code{bbox} class object?
\item \link{is_sf_list}: is x is a list of \code{sf} class objects (with or without names)?
\item \link{is_raster}: is x a \code{Raster} class object?
\item \link{is_sp}: is x a \code{Spatial} class object of any type?
\item \link{is_geo_coords}: is x likely a geodetic coordinate pair (a length 2 numeric vector, with a max absolute value less than or equal to 180)?
\item \link{is_wgs84}: is x using the \href{https://en.wikipedia.org/wiki/World_Geodetic_System}{WSG84} coordinate reference system?
\item \link{is_same_crs}: do x and y have the same coordinate reference system?
}
}
\examples{
nc <- read_sf_ext(system.file("shape/nc.shp", package = "sf"))

is_sf(nc)

is_sfc(nc$geometry)

is_sf_list(list(nc[1, ], nc[2, ]))

is_bbox(sf::st_bbox(nc))

is_wgs84(sf::st_transform(nc, 4326))

is_same_crs(nc, 4267)
}
