% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make file name and path with optional label, prefix, or postfix}
\usage{
make_filename(
  name = NULL,
  label = NULL,
  filetype = NULL,
  filename = NULL,
  path = NULL,
  prefix = NULL,
  postfix = NULL,
  pad = NULL,
  width = NULL,
  cache = FALSE,
  create = TRUE
)
}
\arguments{
\item{name}{Name to make file name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}, e.g. "Residential zoning map" becomes
"residential_zoning_map"}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}. The label is designed to identify the area
or other shared characteristics across multiple data files, maps, or plots.}

\item{filetype}{File type or extension.}

\item{filename}{File name; if file name is \code{NULL}, name and file type are
both required.}

\item{path}{Path to file or data directory.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{width}{Minimum width of padded strings.}

\item{cache}{If \code{TRUE}, path is set to the package cache directory using
\code{\link[=get_data_dir]{get_data_dir()}}; defaults to \code{FALSE}.}

\item{create}{If \code{FALSE} and path does not exist, return path with a warning.
If \code{TRUE} and \link[rlang:is_interactive]{rlang::is_interactive} is \code{TRUE}, ask user if directory
should be created. If the session not interactive and create is \code{TRUE}, a
new directory will be created.}
}
\description{
A helper function to create consistent file names for plots created with data
files exported with \code{\link[=write_sf_ext]{write_sf_ext()}}.
}
\seealso{
Other read_write: 
\code{\link{read_sf_exif}()},
\code{\link{read_sf_ext}()}
}
\concept{read_write}
