% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_markers.R
\name{make_markers}
\alias{make_markers}
\title{Make map markers from a simple feature object}
\usage{
make_markers(
  data,
  groupname_col = NULL,
  group_meta = NULL,
  join = sf::st_intersects,
  geo = FALSE,
  coords = c("lon", "lat"),
  address = "address",
  point = TRUE,
  crs = NULL,
  fn = NULL,
  ...
)
}
\arguments{
\item{data}{Data with markers, passed to data parameter of
\code{\link[=get_location_data]{get_location_data()}}}

\item{groupname_col}{Group column name, used to join group metadata if
group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect
with markers; Default: \code{NULL}}

\item{join}{The join function used by \code{\link[sf:st_join]{sf::st_join()}} if group_meta is
an sf object, Default: \link[sf:geos_binary_pred]{sf::st_intersects}}

\item{geo}{If \code{TRUE}, geocode data using \link[tidygeocoder:geo]{tidygeocoder::geo} and \link{df_to_sf},
Default: \code{FALSE}}

\item{address}{Address column, used if geo is \code{TRUE} Default:
'address'}

\item{point}{If \code{TRUE}, convert geometry to "POINT" with \code{\link[sf:geos_unary]{sf::st_centroid()}},
Default: \code{TRUE}}

\item{crs}{Coordinate reference system for markers, Default: \code{NULL}}

\item{fn}{Function to apply to data before results; gives warning if data is
grouped; Default: \code{NULL}}

\item{...}{Additional parameters passed to \code{\link[=get_location_data]{get_location_data()}}}
}
\description{
Make map markers from a simple feature object
}
