% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_features.R
\name{number_features}
\alias{number_features}
\alias{number_sf}
\alias{sort_features}
\alias{sort_sf}
\title{Sort and number features by coordinates or distance}
\usage{
number_features(
  x,
  col = NULL,
  sort = "dist_xmin_ymax",
  to = NULL,
  desc = FALSE,
  crs = NULL,
  num_style = "arabic",
  num_start = 1,
  suffix = NULL,
  .id = "number"
)

number_sf(
  x,
  col = NULL,
  sort = "dist_xmin_ymax",
  to = NULL,
  desc = FALSE,
  crs = NULL,
  num_style = "arabic",
  num_start = 1,
  suffix = NULL,
  .id = "number"
)

sort_features(
  x,
  col = NULL,
  sort = c("lon", "lat"),
  to = NULL,
  desc = FALSE,
  crs = NULL
)

sort_sf(
  x,
  col = NULL,
  sort = c("lon", "lat"),
  to = NULL,
  desc = FALSE,
  crs = NULL
)
}
\arguments{
\item{x}{A \code{sf} or \code{sfc} object.}

\item{col}{Group column name, Default: \code{NULL}}

\item{sort}{Sort column name, Default: "dist_xmin_ymax".}

\item{to}{A \code{sf} object used to determine sort order based on distance from a
feature to the center of the "to" object.}

\item{desc}{If \code{TRUE}, sort descending; default \code{FALSE}.}

\item{crs}{Coordinate reference to use with \code{\link[=get_coords]{get_coords()}} or \code{\link[=get_minmax]{get_minmax()}}
if "sort" any of the following: "lon", "lat", "longitude", "latitude",
"xmin", "ymin", "xmax", "ymax"}

\item{num_style}{Style of enumeration, either "arabic", "alph", "Alph",
"roman", "Roman".}

\item{num_start}{Starting number; defaults to 1.}

\item{suffix}{Character to appended to "number" column. (e.g. "." for "1." or
":" for "1:"). Can also be a character vector with the same length as the
number column.}

\item{.id}{Name of the column to use for the feature numbers; defaults to
"number".}
}
\value{
A \code{sf} object with a number column ordered by sort values.
}
\description{
Used with \code{\link[maplayer:layer_markers]{maplayer::layer_numbers()}}. Supports multiple types of sorting including
sorting:
\itemize{
\item by centroid coordinates ("lon", "lat") appended with \code{\link[=get_coords]{get_coords()}}
\item by one or more bounding box min or max values ("xmin", "ymin", "xmax",
"ymax") appended with \code{\link[=get_minmax]{get_minmax()}}
\item by distance from the corner, side midpoint, or center of a bounding box
("dist_xmin_ymin", "dist_xmax_ymax", "dist_xmin_ymax", "dist_xmax_ymin",
"dist_xmin_ymid", "dist_xmax_ymid", "dist_xmid_ymin", "dist_xmid_ymax",
"dist_xmid_ymid")
\item by distance to a point (or \code{sf}, \code{sfc}, or \code{bbox} object) passed to the to
parameter
}

For example, in the eastern United States, you can sort and number features
from the top-left corner of the map to the bottom right by setting sort to
"dist_xmin_ymax" (default).

\link{number_features} also supports a range of different numbering styles
designed to match the standard enumeration options available in LaTeX.
}
