% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paper_sizes}
\alias{paper_sizes}
\title{Standard paper and image sizes}
\format{
A data frame with 85 rows and 9 variables:
\describe{
\item{\code{name}}{Name of paper}
\item{\code{series}}{Series}
\item{\code{standard}}{Standard}
\item{\code{size}}{Size in series}
\item{\code{units}}{Units ("in", "mm", or "px") for dimensions}
\item{\code{width}}{Width in units}
\item{\code{height}}{Height in units}
\item{\code{orientation}}{Portrait (width less than height), landscape, or square}
\item{\code{type}}{Type (paper, postcard, print, or social)}
}
}
\usage{
paper_sizes
}
\description{
Reference table of standard paper, postcard, photo print, and social media image
sizes, for \link{get_paper} and \link{get_social_image} functions. Derived from
\href{https://github.com/visioguy/PaperSizes/}{visioguy/PaperSizes} repo, \href{https://www.adobe.com/uk/creativecloud/photography/discover/standard-photo-sizes.html}{Adobe UK guide to photo sizes}
and other sources.
}
\keyword{datasets}
