% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeck_edit.R
\name{rdeck_edit}
\alias{rdeck_edit}
\alias{rdeck_select}
\alias{rdeck_editor_options}
\alias{editor_options}
\title{rdeck editor}
\usage{
rdeck_edit(features, mode = rdeck::cur_value(), initial_bounds = NULL, ...)

rdeck_select(features, ..., mode = "select")

editor_options(mode = rdeck::cur_value(), features = rdeck::cur_value())
}
\arguments{
\item{features}{<\code{\link[rdeck]{wk-geometry}}> Features with which to initialise the editor.
Requires CRS \href{http://epsg.io/4326}{EPSG:4326}.}

\item{mode}{<\code{editor-mode}> The polygon editor mode. One of:
\itemize{
\item \code{view}: editor is in readonly mode
\item \code{select}: select/unselect features
\item \code{modify}: add/move/delete vertices
\item \code{transform}: move/scale/rotate selected features
\item \code{point}: draw points
\item \code{linestring}: draw linestrings by clicking each vertex
\item \code{polygon}: draw polygons by clicking each vertex
\item \code{lasso}: freehand polygon draw by click-dragging
}}

\item{initial_bounds}{<\code{\link[wk:rct]{rct}}/\code{\link[sf:st_bbox]{st_bbox}}/\code{\link[rdeck]{wk-geometry}}>
Sets the initial bounds of the map if not \code{NULL}. Takes priority over \code{initial_view_state}.
Accepts a bounding box, or a geometry from which a bounding box can be computed. Requires
CRS \href{http://epsg.io/4326}{EPSG:4326}.}

\item{...}{
  Arguments passed on to \code{\link[rdeck:rdeck]{rdeck::rdeck}}
  \describe{
    \item{\code{map_style}}{<\code{string}> The mapbox basemap style url.
See \url{https://docs.mapbox.com/api/maps/#mapbox-styles}}
    \item{\code{theme}}{<\code{"kepler"} | \code{"light"}> The widget theme which alters the style of the
legend and tooltips.}
    \item{\code{initial_view_state}}{<\code{\link[rdeck]{view_state}}> Defines the map position, zoom, bearing and pitch.}
    \item{\code{controller}}{<\code{logical}> If \code{NULL} or \code{FALSE}, the map is not interactive.}
    \item{\code{picking_radius}}{<\code{number}> Extra pixels around the pointer to include while picking;
useful when rendering objects that are difficult to hover, e.g. thin lines, small points, etc.}
    \item{\code{use_device_pixels}}{<\code{logical} | \code{number}> Controls the resolution of drawing buffer used
for rendering.
\itemize{
\item \code{TRUE}: Resolution is defined by \code{window.devicePixelRatio}. On Retina/HD displays, this
resolution is usually twice as big as \verb{CSS pixels} resolution.
\item \code{FALSE}: \verb{CSS pixels} resolution is used for rendering.
\item \code{number}: Custom ratio (drawing buffer resolution to \verb{CSS pixel}) to determine drawing
buffer size. A value less than \code{1} uses resolution smaller than \verb{CSS pixels}, improving
rendering performance at the expense of image quality; a value greater than \code{1} improves
image quality at the expense of rendering performance.
}}
    \item{\code{blending_mode}}{<\code{"normal"} | \code{"additive"} | \code{"subtractive"}> Sets the blending mode.
Blending modes:
\itemize{
\item \code{normal}: Normal blending doesn't alter colours of overlapping objects.
\item \code{additive}: Additive blending adds colours of overlapping objects. Useful for highlighting
dot density on dark maps.
\item \code{subtractive}: Subtractive blending darkens overlapping objects. Useful for highlighting
dot density on light maps.
}}
    \item{\code{layer_selector}}{<\code{boolean}> If \code{TRUE}, the layer selector control will be enabled
and layers with \code{visibility_toggle = TRUE} may be toggled. If \code{FALSE}, the layer selector control
won't be rendered.}
    \item{\code{editor}}{<\code{boolean}|\code{\link[rdeck]{editor_options}}> Whether to render the polygon editor.
If \code{TRUE}, renders with the default \code{\link[rdeck:editor_options]{editor_options()}}. If \code{FALSE}, the polygon editor
is not rendered.}
    \item{\code{lazy_load}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Maps are always eagerly rendered.}
    \item{\code{width}}{<\code{number}> The width of the map canvas.}
    \item{\code{height}}{<\code{number}> The height of the map canvas.}
    \item{\code{id}}{<\code{string}> The map element id. Not used in shiny applications.}
  }}
}
\description{
A wrapper for \code{\link[rdeck:editor_options]{rdeck::editor_options()}} that automatically converts features
to WGS84.
}
