% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ext.R
\name{read_sf_ext}
\alias{read_sf_ext}
\alias{read_sf_pkg}
\alias{read_sf_path}
\alias{read_sf_zip}
\alias{read_sf_rdata}
\alias{read_sf_query}
\alias{read_sf_excel}
\alias{read_sf_csv}
\alias{read_sf_url}
\alias{read_sf_esri}
\alias{read_sf_gist}
\alias{read_sf_gmap}
\alias{read_sf_download}
\alias{read_sf_gsheet}
\title{Read spatial data in a bounding box to a simple feature object from multiple
sources}
\usage{
read_sf_ext(...)

read_sf_pkg(
  data,
  bbox = NULL,
  package = NULL,
  pkg = NULL,
  fileext = "gpkg",
  filetype = NULL,
  ...
)

read_sf_path(path, bbox = NULL, ...)

read_sf_zip(
  path,
  bbox = NULL,
  exdir = NULL,
  overwrite = TRUE,
  unzip = "internal",
  ...
)

read_sf_rdata(
  path,
  file = NULL,
  refhook = NULL,
  bbox = NULL,
  .name_repair = "check_unique",
  ...
)

read_sf_query(
  path,
  dsn = NULL,
  bbox = NULL,
  query = NULL,
  table = NULL,
  name = NULL,
  name_col = NULL,
  wkt_filter = NULL,
  zm_drop = FALSE,
  .name_repair = "check_unique",
  ...
)

read_sf_excel(
  path,
  sheet = NULL,
  combine_sheets = FALSE,
  bbox = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  .name_repair = "check_unique",
  ...
)

read_sf_csv(
  path,
  url = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  wkt = NULL,
  .name_repair = "check_unique",
  show_col_types = FALSE,
  ...
)

read_sf_url(url, bbox = NULL, coords = c("lon", "lat"), ...)

read_sf_esri(
  url,
  bbox = NULL,
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  .name_repair = "check_unique",
  ...
)

read_sf_gist(url, id = NULL, bbox = NULL, nth = 1, ...)

read_sf_gmap(
  url,
  bbox = NULL,
  layer = NULL,
  combine_layers = FALSE,
  zm_drop = TRUE,
  .name_repair = "check_unique"
)

read_sf_download(
  url,
  filename,
  bbox = NULL,
  path = NULL,
  filetype = "geojson",
  prefix = "date",
  method = "auto",
  unzip = FALSE,
  .name_repair = "check_unique",
  ...
)

read_sf_gsheet(
  url,
  sheet = NULL,
  ss = NULL,
  bbox = NULL,
  ask = FALSE,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  .name_repair = "check_unique",
  ...
)
}
\arguments{
\item{...}{Additional parameters passed to multiple functions; see details.}

\item{data}{Name of a package dataset; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only.}

\item{bbox}{A bounding box object; defaults to \code{NULL}. If \code{"bbox"} is
provided, only returns features intersecting the bounding box.}

\item{package, pkg}{Package name; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only. pkg is used if
package is \code{NULL}.}

\item{fileext}{File extension. If supplied to \code{\link[filenamr:list_path_filenames]{list_path_filenames()}} and
pattern is \code{NULL}, only return file names matching this extension.}

\item{filetype}{File type supported by \code{\link[sf:read_sf]{sf::read_sf()}}; Default: 'gpkg'; used
by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only and required only if the data is in the package
cache directory or extdata system files.}

\item{path}{A file path.}

\item{exdir}{The directory to extract files to (the equivalent of
    \code{unzip -d}).  It will be created if necessary.}

\item{overwrite}{If \code{TRUE}, overwrite existing files (the equivalent
    of \command{unzip -o}), otherwise ignore such files (the equivalent of
    \command{unzip -n}).}

\item{unzip}{If \code{TRUE}, url must be a zip file that is downloaded to a cache
folder, unzipped into a temporary directory (created with \code{\link[=tempdir]{tempdir()}}), and
then read to a file using the specified file type.}

\item{file}{The file path to read from/write to.}

\item{refhook}{A function to handle reference objects.}

\item{.name_repair}{Passed to repair parameter of \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}}

\item{dsn}{data source name (interpretation varies by driver - for some
drivers, \code{dsn} is a file name, but may also be a folder, or contain
the name and access credentials of a database); in case of GeoJSON,
\code{dsn} may be the character string holding the geojson data. It can
also be an open database connection.}

\item{query}{SQL query to select records; see details}

\item{table}{table can usually be inferred from basename of the data source.
table is used to generate a custom query if both name and name_col are
provided. Use \code{sf::st_layers(dsn = dsn)[["name"]]} to see a list of
available table names.}

\item{name, name_col}{Name value and name column to use in generated a query
for sources read with \code{\link[=read_sf_query]{read_sf_query()}} or \code{\link[=read_sf_esri]{read_sf_esri()}}.}

\item{wkt_filter}{character; WKT representation of a spatial filter (may be used as bounding box, selecting overlapping geometries); see examples}

\item{zm_drop}{If \code{TRUE}, drop Z and/or M dimensions using \link[sf:st_zm]{sf::st_zm}}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{coords}{Character vector with coordinate values. Coordinates must use
the same crs as the \code{from_crs} parameter.}

\item{from_crs}{For \code{\link[=df_to_sf]{df_to_sf()}}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{geo}{If \code{TRUE}, use \code{\link[=address_to_sf]{address_to_sf()}} to geocode address column;
defaults to \code{FALSE}.}

\item{address}{Address column name, Default: 'address'}

\item{url}{A url for a spatial data file, tabular data with coordinates, or a
ArcGIS FeatureServer or MapServer to access with \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf()}}}

\item{wkt}{Name of column with well-known text for geometry. Used by
\code{\link[=read_sf_csv]{read_sf_csv()}}.}

\item{show_col_types}{If \code{FALSE}, do not show the guessed column types. If
\code{TRUE} always show the column types, even if they are supplied. If \code{NULL}
(the default) only show the column types if they are not explicitly supplied
by the \code{col_types} argument.}

\item{where}{A where condition (SQL-like syntax) passed to
\code{\link[esri2sf:esri2sf]{esri2sf::esri2sf()}} to filter features returned from an ArcGIS
FeatureServer or MapServer, e.g. \code{"POP2000 > 100000"}. Defaults to \code{NULL}.}

\item{id}{The name of a column in which to store the file path. This is
useful when reading multiple input files and there is data in the file
paths, such as the data collection date. If \code{NULL} (the default) no extra
column is created.}

\item{nth}{For \code{\link[=read_sf_gist]{read_sf_gist()}}, the file to return from the gist, e.g. 1 for
first, 2 for second. Defaults to 1.}

\item{layer}{layer name (varies by driver, may be a file name without
extension); in case \code{layer} is missing, \code{st_read} will read the
first layer of \code{dsn}, give a warning and (unless \code{quiet = TRUE})
print a message when there are multiple layers, or give an error if there
are no layers in \code{dsn}. If \code{dsn} is a database connection, then
\code{layer} can be a table name or a database identifier (see
\code{\link[DBI]{Id}}). It is also possible to omit \code{layer} and rather
use the \code{query} argument.}

\item{combine_layers, combine_sheets}{If \code{FALSE} (default), return a list with
a sf object for each layer or sheet as a separate item. If \code{TRUE}, use
\code{\link[purrr:map_dfr]{purrr::map_dfr()}} to combine layers or sheets into a single sf object
using the layer or sheet name as an additional column.}

\item{filename}{File name; if filename is \code{NULL} and path does not include a
file extension, name and file extension are both required.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{method}{Method to be used for downloading files.  Current
    download methods are \code{"internal"}, \code{"libcurl"},
    \code{"wget"}, \code{"curl"} and \code{"wininet"} (Windows
    only), and there is a value \code{"auto"}: see \sQuote{Details} and
    \sQuote{Note}.

    The method can also be set through the option
    \code{"download.file.method"}: see \code{\link{options}()}.
  }

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{ask}{If \code{TRUE}, ask for the name of the Google Sheet to read if ss is
not provided to \link{read_sf_gsheet}.}
}
\description{
An extended version of \code{\link[sf:read_sf]{sf::read_sf()}} that support reading spatial data
based on a file path, URL, or the data name and associated package. A RDS,
RDA, or RData file Optionally provide a bounding box to filter data (data is
filtered before download or reading into memory where possible).
}
\details{
Reading data from a url:

\code{\link[=read_sf_url]{read_sf_url()}} supports multiple types of urls:
\itemize{
\item A MapServer or FeatureServer URL
\item A URL for a GitHub gist with a single spatial data file (first file used
if gist contains multiple)
\item A URL for a spatial data file, CSV file, Excel file, or RDS file (RDA and
RData files supported by \code{\link[=read_sf_path]{read_sf_path()}})
\item A Google Sheets URL
\item A public Google Maps URL
}
}
\section{Reading data from a package}{


\code{\link[=read_sf_pkg]{read_sf_pkg()}} looks for three types of package data:
\itemize{
\item Data loaded with the package
\item External data in the \code{extdata} system files folder.
\item Cached data in the cache directory returned by
\code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}
}
}

\section{Additional ... parameters}{


\code{\link[=read_sf_ext]{read_sf_ext()}} is a flexible function where ... are passed to one of the
other read functions depending on the provided parameters. If you are using
more than one parameter, all parameters \emph{must} be named.

\code{\link[=read_sf_pkg]{read_sf_pkg()}} and \code{\link[=read_sf_download]{read_sf_download()}} both pass additional parameters to
\code{\link[=read_sf_path]{read_sf_path()}} which supports query, name_col, name, and table. name and
name_col are ignored if a query parameter is provided. If table is not
provided, a expected layer name is created based on the file path.

\code{\link[=read_sf_url]{read_sf_url()}} pass the where, name_col, and name for any ArcGIS
FeatureServer or MapServer url (passed to \code{\link[=read_sf_esri]{read_sf_esri()}}) or sheet if the
url is for a Google Sheet (passed to \code{\link[googlesheets4:read_sheet]{googlesheets4::read_sheet()}}), or a
query or wkt filter parameter if the url is some other type (passed to
\code{\link[sf:read_sf]{sf::read_sf()}}).
}

\seealso{
Other read_write:
\code{\link[=read_sf_exif]{read_sf_exif()}}
}
\concept{read_write}
