% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox_corners.R
\name{sf_bbox_corners}
\alias{sf_bbox_corners}
\title{Get bounding box corner points from a bbox, sfc, or sf object}
\usage{
sf_bbox_corners(x, .id = "corner")
}
\arguments{
\item{x}{A \code{bbox}, length 1 \code{sfc} object, or a \code{sf} object with a single
feature}

\item{.id}{A name to use for added column with cardinal position of points,
Default: 'corner'}
}
\value{
A \code{sf} with four features or a length 4 \code{sfc} object with POINT
geometry.
}
\description{
If x is an sf object, the results include a column with a name matching .id
containing the values "SW", "SE", "NE", and "NW" corresponding to the
cardinal position of each corner.
}
