% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox_dist.R
\name{sf_bbox_dist}
\alias{sf_bbox_dist}
\alias{sf_bbox_xdist}
\alias{sf_bbox_ydist}
\alias{sf_bbox_diagdist}
\alias{sf_bbox_diag_ratio_to_dist}
\alias{sf_bbox_asp}
\alias{sf_bbox_orientation}
\alias{sf_bbox_check_fit}
\title{Measure a bounding box using x, y, or diagonal distance}
\usage{
sf_bbox_dist(
  bbox,
  from,
  to,
  units = NULL,
  drop = TRUE,
  by_element = TRUE,
  call = caller_env(),
  ...
)

sf_bbox_xdist(bbox, units = NULL, drop = TRUE)

sf_bbox_ydist(bbox, units = NULL, drop = TRUE)

sf_bbox_diagdist(bbox, units = NULL, drop = TRUE)

sf_bbox_diag_ratio_to_dist(bbox, diag_ratio, units = NULL, drop = TRUE)

sf_bbox_asp(bbox)

sf_bbox_orientation(bbox, tolerance = 0.1, call = caller_env())

sf_bbox_check_fit(bbox, dist)
}
\arguments{
\item{bbox}{A \code{bbox} object.}

\item{from, to}{Length 2 character vectors with xy pairs (e.g. c("xmax",
"ymax) defining points to measure distance from and to.}

\item{units}{The units to return for sf_bbox_dist. Defaults to NULL.}

\item{drop}{If \code{FALSE}, distance functions return with units. If \code{FALSE}
(default), distance functions return numeric values.}

\item{by_element}{logical; if \code{TRUE}, return a vector with distance between the first elements of \code{x} and \code{y}, the second, etc; an error is raised if \code{x} and \code{y} are not the same length. If \code{FALSE}, return the dense matrix with all pairwise distances.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{...}{passed on to \link[s2]{s2_distance}, \link[s2]{s2_distance_matrix}, or \link[s2]{s2_perimeter}}

\item{diag_ratio}{Proportion of the diagonal distance (ratio of distance to
the full diagonal distance) across the bounding box.}

\item{tolerance}{Numeric value above or below 1 used by
\code{\link[=sf_bbox_orientation]{sf_bbox_orientation()}} to describe an aspect ratio as landscape or
portrait; defaults to 0.1.}

\item{dist}{Distance to compare with bounding box for \code{\link[=sf_bbox_check_fit]{sf_bbox_check_fit()}}.
If dist is length 1 compare to bounding box diagonal distance. If dist is
length 2, compare to bounding box x and y distances. Return TRUE if dist
fits within bounding box or FALSE if dist exceeds bounding box limits.}
}
\description{
\itemize{
\item Measure distances (\code{\link[=sf_bbox_dist]{sf_bbox_dist()}}, \code{\link[=sf_bbox_xdist]{sf_bbox_xdist()}}, \code{\link[=sf_bbox_ydist]{sf_bbox_ydist()}},
and \code{\link[=sf_bbox_diagdist]{sf_bbox_diagdist()}})
\item Convert a diag_ratio value to a distance value if a bbox and diag_ratio are
provided (\code{\link[=sf_bbox_diag_ratio_to_dist]{sf_bbox_diag_ratio_to_dist()}}); return \code{NULL} if bbox and
diag_ratio are \code{NULL}
\item Get an aspect ratio or orientation (\code{\link[=sf_bbox_asp]{sf_bbox_asp()}}) (counts asp between
0.9 and 1.1 as "square" when tolerance is 0.1) or (\code{\link[=sf_bbox_orientation]{sf_bbox_orientation()}})
}
}
\seealso{
Other dist:
\code{\link[=compare_dist]{compare_dist()}},
\code{\link[=convert_dist_scale]{convert_dist_scale()}},
\code{\link[=convert_dist_units]{convert_dist_units()}},
\code{\link{get_measurements}},
\code{\link[=is_dist_units]{is_dist_units()}}
}
\concept{dist}
