% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox_shift.R
\name{sf_bbox_shift}
\alias{sf_bbox_shift}
\alias{sf_bbox_contract}
\alias{sf_bbox_expand}
\title{Shift sides, contract, or expand a bounding box}
\usage{
sf_bbox_shift(
  bbox,
  nudge_x = 0,
  nudge_y = 0,
  side = c("all", "top", "bottom", "left", "right"),
  dir = NULL,
  call = caller_env()
)

sf_bbox_contract(bbox, nudge_x = 0, nudge_y = 0)

sf_bbox_expand(bbox, nudge_x = 0, nudge_y = 0)
}
\arguments{
\item{bbox}{A \code{bbox} object.}

\item{nudge_x, nudge_y}{Length 1 or 2 numeric vector; unitless. Used by or
passed to \code{\link[=sf_bbox_shift]{sf_bbox_shift()}}. Required for  \code{\link[=sf_bbox_shift]{sf_bbox_shift()}}.}

\item{side}{one or more sides to shift: "top", "bottom", "left", "right", or
"all". Required for \code{\link[=sf_bbox_shift]{sf_bbox_shift()}}.}

\item{dir}{If "in", contract the \code{bbox} by nudge_x and nudge_y. If "out",
expand the bbox by nudge_x and nudge_y. If dir is not \code{NULL}; absolute
values are used for nudge_x and nudge_y. Defaults to \code{NULL}. Optional
\code{\link[=sf_bbox_shift]{sf_bbox_shift()}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\itemize{
\item Shift, expand, or contract a bounding box (\code{\link[=sf_bbox_shift]{sf_bbox_shift()}},
\code{\link[=sf_bbox_expand]{sf_bbox_expand()}}, \code{\link[=sf_bbox_contract]{sf_bbox_contract()}})
}
}
