% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfext-package.R
\docType{package}
\name{sfext-package}
\alias{sfext}
\alias{sfext-package}
\title{sfext: Extra Functions for Simple Feature Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extra functions with additional options for reading, writing, and transforming spatial data. Includes a variety of utility functions for working with tabular data with coordinates and distance and area units.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/sfext}
  \item \url{https://elipousson.github.io/sfext/}
  \item Report bugs at \url{https://github.com/elipousson/sfext/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
