% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_ext.R
\name{st_bbox_ext}
\alias{st_bbox_ext}
\alias{st_bbox_adj}
\alias{st_bbox_ext.default}
\alias{st_bbox_ext.bbox}
\alias{st_bbox_ext.list}
\alias{st_bbox_ext.sf_list}
\alias{st_bbox_asp}
\alias{st_bbox_asp.default}
\alias{st_bbox_asp.bbox}
\alias{st_bbox_asp.sf_list}
\alias{st_bbox_asp.list}
\title{Get a bounding box buffered a set distance or to match an aspect ratio}
\usage{
st_bbox_ext(
  x,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = NULL,
  crs = NULL,
  class = "bbox",
  nudge = NULL,
  allow_null = TRUE,
  allow_list = TRUE
)

\method{st_bbox_ext}{default}(
  x,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = NULL,
  crs = NULL,
  class = "bbox",
  nudge = NULL,
  allow_null = TRUE,
  ...
)

\method{st_bbox_ext}{bbox}(
  x,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = NULL,
  crs = NULL,
  class = "bbox",
  nudge = NULL,
  allow_null = TRUE,
  ...
)

\method{st_bbox_ext}{list}(
  x,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = NULL,
  crs = NULL,
  class = "bbox",
  nudge = NULL,
  allow_null = TRUE,
  allow_list = TRUE,
  ...
)

\method{st_bbox_ext}{sf_list}(
  x,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = NULL,
  crs = NULL,
  class = "bbox",
  nudge = NULL,
  allow_null = TRUE,
  allow_list = TRUE,
  ...
)

st_bbox_asp(
  x,
  asp = NULL,
  class = "bbox",
  allow_null = TRUE,
  allow_list = TRUE
)

\method{st_bbox_asp}{default}(x, asp = NULL, class = "bbox", allow_null = TRUE)

\method{st_bbox_asp}{bbox}(x, asp = NULL, class = "bbox")

\method{st_bbox_asp}{sf_list}(
  x,
  asp = NULL,
  class = "bbox",
  allow_null = TRUE,
  allow_list = TRUE
)

\method{st_bbox_asp}{list}(
  x,
  asp = NULL,
  class = "bbox",
  allow_null = TRUE,
  allow_list = TRUE
)
}
\arguments{
\item{x}{An object \code{sf}, \code{bbox}, \code{sfc}, \code{raster}, or \code{sp} object or a
\code{data.frame} that can be converted to an \code{sf} object or a list of \code{sf},
\code{bbox}, or \code{sfc} objects. \code{\link[=st_bbox_asp]{st_bbox_asp()}} also supports vectors in the same
format as a \code{bbox} object.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{crs}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}

\item{class}{Class of object to return (\code{sf} or \code{bbox}); defaults to "bbox".}

\item{nudge}{Passed as to parameter \code{\link[=st_nudge]{st_nudge()}} when not \code{NULL}. Numeric
vector or sf object.}

\item{allow_null}{If \code{TRUE} and x is \code{NULL}, return \code{NULL}.}

\item{allow_list}{If \code{TRUE}, allow sf list objects as an input and use
\code{\link[purrr:map]{purrr::map()}} to apply the provided parameters to each object within the
list to return as a new sf list object.}

\item{...}{Additional parameters passed to \code{\link[sf:geos_unary]{sf::st_buffer()}}}
}
\value{
A \code{bbox} object converted to match the class of the class parameter.
}
\description{
This function takes a \code{sf} object and returns a bounding box buffered by a
specified distance or diagonal ratio (a proportion of the diagonal distance
across the bounding box) and to match the provided aspect ratio.
}
\details{
Common aspect ratios include "1:1" (1), "4:6" (0.666), "8.5:11", "16:9"
(1.777). The asp parameter supports both numeric values and character
strings with ratios matching the format of "width:height".
}
