% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_cast_ext.R
\name{st_cast_ext}
\alias{st_cast_ext}
\title{Cast geometry of a simple feature object or simple feature collection to
another type}
\usage{
st_cast_ext(x, to = "POINT", simplify = TRUE, ...)
}
\arguments{
\item{x}{A \code{sf} or \code{sfc} object to cast to another type.}

\item{to}{character; target type, if missing, simplification is tried; when \code{x} is of type \code{sfg} (i.e., a single geometry) then \code{to} needs to be specified.}

\item{simplify}{If \code{TRUE}, simplify geometry type; defaults to \code{TRUE}.}

\item{...}{
  Arguments passed on to \code{\link[sf:st_cast]{sf::st_cast}}
  \describe{
    \item{\code{ids}}{integer vector, denoting how geometries should be grouped (default: no grouping)}
    \item{\code{group_or_split}}{logical; if TRUE, group or split geometries; if FALSE, carry out a 1-1 per-geometry conversion.}
    \item{\code{warn}}{logical; if \code{TRUE}, warn if attributes are assigned to sub-geometries}
    \item{\code{do_split}}{logical; if \code{TRUE}, allow splitting of geometries in sub-geometries}
  }}
}
\description{
Wrapper for \code{\link[sf:st_cast]{sf::st_cast()}} that currently supports casting MULTIPOLYGON to
POLYGON or MULTIPOLYGON or POLYGON to POINT or, if simplify = FALSE, can cast
MULTIPOINT to LINESTRING. This is not very different than the basic
functionality of st_cast but further development may improve the utility of
this function.
}
