% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_concave_hull.R
\name{st_concave_hull}
\alias{st_concave_hull}
\title{Make a concave hull around simple feature object by attribute}
\usage{
st_concave_hull(
  x,
  by = NULL,
  centroid = FALSE,
  concavity = 2,
  length_threshold = 0
)
}
\arguments{
\item{x}{A sf object.}

\item{by}{Column name to use for grouping and combining geometry, Default:
\code{NULL}}

\item{centroid}{If \code{TRUE}, use centroids for geometry of x, Default: \code{FALSE}}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{length_threshold}{when a segment length is under this threshold, it stops being considered for further detalization. Higher values result in simpler shapes.}
}
\description{
Make a concave hull around simple feature object by attribute
}
\seealso{
\code{\link[concaveman]{concaveman}}
}
