% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_concave_hull.R
\name{st_concave_hull_ext}
\alias{st_concave_hull_ext}
\title{Make a concave hull around simple feature object by attribute}
\usage{
st_concave_hull_ext(
  x,
  by = NULL,
  centroid = FALSE,
  ratio = 0.5,
  allow_holes = FALSE
)
}
\arguments{
\item{x}{A sf object.}

\item{by}{Column name to use for grouping and combining geometry, Default:
\code{NULL}}

\item{centroid}{If \code{TRUE}, use centroids for geometry of x, Default: \code{FALSE}}

\item{ratio}{numeric; fraction convex: 1 returns the convex hulls, 0 maximally concave hulls}

\item{allow_holes}{logical; if \code{TRUE}, the resulting concave hull may have holes}
}
\description{
Make a concave hull around simple feature object by attribute
}
