% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_dissolve.R
\name{st_dissolve}
\alias{st_dissolve}
\title{Dissolve geometry preserving existing or supplied grouping variables}
\usage{
st_dissolve(
  x,
  ...,
  .by = NULL,
  .keep = "nest",
  do_union = TRUE,
  .data_key = "data",
  .dissolve_key = "group.comp.id",
  call = caller_env
)
}
\arguments{
\item{x}{A sf or sfc object. If sf object, a grouped data frame is allowed.}

\item{...}{
  Arguments passed on to \code{\link[spdep:poly2nb]{spdep::poly2nb}}
  \describe{
    \item{\code{pl}}{list of polygons of class extending \code{SpatialPolygons}, or an \code{sf} or \code{sfc} object containing non-empty (multi-)polygon objects}
    \item{\code{row.names}}{character vector of region ids to be added to the neighbours list as attribute \code{region.id}, default \code{seq(1, nrow(x))}; if \code{pl} has \code{row.names}, they are used instead of the default sequence.}
    \item{\code{snap}}{boundary points less than \code{snap} distance apart are considered to indicate contiguity; used both to find candidate and actual neighbours for planar geometries, but only actual neighbours for spherical geometries, as spherical spatial indexing itself injects some fuzzyness. If not set, for all \code{SpatialPolygons} objects, the default is as before \code{sqrt(.Machine$double.eps)}, with this value also used for \code{sf} objects with no coordinate reference system. For \code{sf} objects with a defined coordinate reference system, the default value is \code{1e-7} for geographical coordinates (approximately 10mm), is 10mm where projected coordinates are in metre units, and is converted from 10mm to the same distance in the units of the coordinates. Should the conversion fail, \code{snap} reverts to \code{sqrt(.Machine$double.eps)}.}
    \item{\code{queen}}{if TRUE, a single shared boundary point meets the contiguity condition, if FALSE, more than one shared point is required; note that more than one shared boundary point does not necessarily mean a shared boundary line}
    \item{\code{useC}}{default TRUE, doing the work loop in C, may be set to false to revert to R code calling two C functions in an \code{n*k} work loop, where \code{k} is the average number of candidate neighbours}
    \item{\code{foundInBox}}{default NULL using R code or \code{st_intersects()} to generate candidate neighbours (using \code{snap=} if the geometries are not spherical); if not NULL (for legacy purposes) a list of length \code{(n-1)} with integer vectors of candidate neighbours \code{(j > i)} (as created by the \code{poly_findInBoxGEOS} function in \pkg{rgeos} for clean polygons)}
  }}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}

\item{.keep}{Method for handling attributes for input data. By default
("nest"), input data is converted to a nested list column with the name
from \code{.data_key}. Any other values}

\item{do_union}{Use \code{\link[sf:geos_combine]{sf::st_union()}} to dissolve geometry if \code{TRUE}
(default). Use \code{\link[sf:geos_combine]{sf::st_combine()}} if \code{FALSE}.}

\item{.data_key}{Passed to \code{.data} argument of \code{\link[tidyr:nest]{tidyr::nest()}}}

\item{.dissolve_key}{Used as the column name for the dissolve grouping
variable. x can't have any existing columns with this name.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=st_dissolve]{st_dissolve()}} dissolves sf and sfc objects. If the input object is an sf
object, any existing grouping variables or added grouping variables passed to
the \code{.by} parameter are included in the output sf data frame.
}
