% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter_ext.R
\name{st_filter_ext}
\alias{st_filter_ext}
\alias{st_filter_geom_type}
\title{Filter by geometry type}
\usage{
st_filter_geom_type(x, type = NULL)
}
\arguments{
\item{type}{Geometry type.}
}
\description{
Filter by geometry type
}
