% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter_pct.R
\name{st_filter_pct}
\alias{st_filter_pct}
\alias{st_filter_pct_area}
\alias{st_filter_pct_length}
\title{Filter by share of length or area of one geometry overlapping with a second
geometry}
\usage{
st_filter_pct(x, y, pct = NULL, ...)

st_filter_pct_area(x, y, pct = NULL)

st_filter_pct_length(x, y, pct = NULL)
}
\arguments{
\item{x}{A sf object to filter.}

\item{y}{A sf or sfc object to filter by.}

\item{pct}{Percent of length or area to use as a threshold value for filter.
Numeric value of 1 or less.}

\item{...}{Additional parameters. Not used currently.}
}
\value{
A filtered version of the input sf object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
