% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_intersection_ext.R
\name{st_intersection_ext}
\alias{st_intersection_ext}
\title{A wrapper for \code{sf::st_intersection()}}
\usage{
st_intersection_ext(x, y = NULL, crs = NULL, ...)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{y}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{...}{
  Arguments passed on to \code{\link[sf:geos_binary_ops]{sf::st_intersection}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A sf or sfc object.
}
\description{
Includes checks for empty input for x and y. Also allows the specification of
an output coordinate reference system.
}
\seealso{
\code{\link[sf:geos_binary_ops]{sf::geos_binary_ops()}}
}
