% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_erase.R
\name{st_make_valid_ext}
\alias{st_make_valid_ext}
\title{Checks if all geometries are already valid and make valid if not}
\usage{
st_make_valid_ext(x, ...)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{...}{passed on to \link[s2]{s2_options}}
}
\description{
Checks if all geometries are already valid and make valid if not
}
