% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filename.R
\name{str_misc}
\alias{str_misc}
\alias{str_fix}
\alias{str_prefix}
\alias{str_pad_digits}
\alias{str_extract_digits}
\alias{str_add_filetype}
\alias{str_remove_filetype}
\alias{str_extract_filetype}
\title{Modify strings to help make consistent file names}
\usage{
str_fix(
  prefix = NULL,
  string = NULL,
  postfix = NULL,
  sep = "_",
  clean_names = TRUE,
  pad = NULL,
  width = NULL
)

str_prefix(
  string = NULL,
  prefix = NULL,
  sep = "_",
  clean_names = TRUE,
  post = FALSE,
  date.format = "\%F",
  time.format = "\%Y-\%m-\%d_\%I-\%M-\%S_\%p"
)

str_pad_digits(string, pad = "0", side = "left", width = NULL)

str_extract_digits(string)

str_add_filetype(string, filetype = NULL)

str_remove_filetype(string, filetype = NULL)

str_extract_filetype(string)
}
\arguments{
\item{prefix}{Character string or character vector to add to string parameter
as a prefix.}

\item{string}{A single string that the attach prefix or postfix is added to.}

\item{postfix}{Character string or character vector to add to string
parameter as a postfix.}

\item{sep}{Separator character passed as the collapse parameter of \code{\link[=paste]{paste()}}.}

\item{clean_names}{If \code{TRUE}, prefix, postfix, and string are all converted
to snake case with \code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{width}{Minimum width of padded strings.}

\item{post}{If \code{TRUE}, use the prefix string as a postfix; defaults to
\code{FALSE}.}

\item{date.format, time.format}{Date or time format Only used by \link{str_prefix} if
prefix is "date" or "time" and not currently accessible when using
\link{str_fix} or \link{make_filename}.}

\item{side}{Side on which padding character is added (left, right or both).}

\item{filetype}{File type string}
}
\description{
Prefix and postfix can include more than one value that are added in the same
order provided. For \link{str_fix}, the string must be a single character string.
}
\details{
Functions include:
\itemize{
\item \link{str_fix}: Add a label, prefix, and postfix to string
\item \link{str_prefix}: Add a prefix or a postfix to a string
\item \link{str_add_filetype}: Add file type to string
\item \link{str_remove_filetype}: Remove file type from string
\item \link{str_extract_filetype}: Extract file type from string
\item \link{str_pad_digits}: Pad a string with digits
\item \link{str_extract_digits}: Extract digits from a string
}
}
