% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_sf.R
\name{vec_st_union}
\alias{vec_st_union}
\title{Vectorized wrapper for \code{sf::st_union()}}
\usage{
vec_st_union(x, y, cbind = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{y}{object of class \code{sf}, \code{sfc} or \code{sfg} (optional)}

\item{cbind}{If \code{TRUE}, bind the attributes for y to x assuming y is the same
length as x.}

\item{...}{ignored}
}
\description{
\code{\link[=vec_st_union]{vec_st_union()}} adds vctrs recycling rules to \code{\link[sf:geos_combine]{sf::st_union()}} by assuming
that y must be length 1 or the same length as x and that if y is the same
length as x, the function should union each element in turn, e.g. first
feature of x unioned with first feature of y and so on.
}
\keyword{internal}
