% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_ext.R
\name{write_sf_svg}
\alias{write_sf_svg}
\title{Write an sf object to an svg file}
\usage{
write_sf_svg(data, filename = NULL, path = NULL, ..., width = 10, height = 10)
}
\arguments{
\item{data}{A sf object to save as a svg file.}

\item{filename}{the file path of the output file(s).
    The page number is substituted if a C integer format is included in
    the character string, as in the default.  (Depending on the platform,
    the result must be less
    than \code{PATH_MAX} characters long, and may be truncated if not.
    See \code{\link[grDevices]{pdf}} for further details.)  Tilde expansion
    is performed where supported by the platform.}

\item{path}{File path to combine with filename. Optional if filename is
provided. filename is optional if path includes a svg file extension.}

\item{...}{
  Arguments passed on to \code{\link[grDevices:svg]{grDevices::svg}}
  \describe{
    \item{\code{pointsize}}{the default pointsize of plotted text (in big points).}
    \item{\code{onefile}}{should all plots appear in one file or in separate files?}
    \item{\code{family}}{one of the device-independent font families,
    \code{"sans"}, \code{"serif"} and \code{"mono"}, or a character
    string specify a font family to be searched for in a
    system-dependent way.

    On unix-alikes (incl.\\ macOS), see
    the \sQuote{Cairo fonts} section in the help for \code{\link[grDevices]{X11}}.
  }
    \item{\code{bg}}{the initial background colour: can be overridden by setting
    \code{par("bg")}.}
    \item{\code{antialias}}{string, the type of anti-aliasing (if any) to be used;
    defaults to \code{"default"}.}
    \item{\code{fallback_resolution}}{numeric: the resolution in dpi used when
    falling back to bitmap output.}
    \item{\code{symbolfamily}}{
      a length-one character string that specifies the font family to
      be used as the "symbol" font (e.g., for \link[grDevices]{plotmath} output).}
  }}

\item{width}{the width of the device in inches.}

\item{height}{the height of the device in inches.}
}
\description{
\code{\link[=write_sf_svg]{write_sf_svg()}} uses \code{\link[=plot]{plot()}} and \code{\link[=svg]{svg()}} to create a simple plot of an sf
object geometry. This function is convenient for working with designers or
other collaborators interested in using spatial data outside of R or a
desktop GIS application.
}
