% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_df.R
\name{address_to_sf}
\alias{address_to_sf}
\title{Use tidygeocoder to convert an address or data frame with an address column
to an sf object}
\usage{
address_to_sf(
  x,
  address = "address",
  method = "osm",
  coords = c("lon", "lat"),
  remove_coords = FALSE,
  crs = NULL,
  full_results = FALSE,
  ...,
  call = caller_env()
)
}
\arguments{
\item{x}{Data frame with an address column. Multiple address columns are not
currently supported.}

\item{address}{Address column name, Default: 'address'}

\item{method}{Geocoding service to use, Default: 'osm'. Passed to
\code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}} which passes it on to \code{\link[tidygeocoder:geo]{tidygeocoder::geo()}}.
See \code{\link[tidygeocoder:geo]{tidygeocoder::geo()}} for supported services (e.g. "osm", "census",
"arcgis", "google").}

\item{coords}{Coordinate columns for input data.frame or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{remove_coords}{For \code{\link[=df_to_sf]{df_to_sf()}}, if \code{TRUE}, remove the coordinate columns
after converting a data frame to simple feature object; defaults to
\code{FALSE}.}

\item{crs}{Cordinate reference system to return, Default: 4326 for
\code{\link[=sf_to_df]{sf_to_df()}} and \code{NULL} for \code{\link[=df_to_sf]{df_to_sf()}}.}

\item{full_results}{If \code{TRUE}, return all columns from the geocoding
service in addition to the latitude and longitude columns, Default:
\code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[tidygeocoder:geocode]{tidygeocoder::geocode}}
  \describe{
    \item{\code{.tbl}}{dataframe containing addresses}
    \item{\code{street}}{street address column name}
    \item{\code{city}}{city column name}
    \item{\code{county}}{county column name}
    \item{\code{state}}{state column name}
    \item{\code{postalcode}}{postal code column name (zip code if in the United States)}
    \item{\code{country}}{country column name}
    \item{\code{lat}}{latitude column name. Can be quoted or unquoted (ie. \code{lat} or \code{"lat"}).}
    \item{\code{long}}{longitude column name. Can be quoted or unquoted (ie. \code{long} or \code{"long"}).}
    \item{\code{return_input}}{if TRUE then the input dataset will be combined with the geocoder query results
and returned. If FALSE only the geocoder results will be returned.}
    \item{\code{limit}}{maximum number of results to return per input address. For many geocoding services the maximum value of the limit parameter is 100. Pass \code{limit = NULL} to use the default \code{limit} value of the selected geocoding service. For batch geocoding, limit must be set to 1 (default) if \code{return_addresses = TRUE}.To use \code{limit > 1} or \code{limit = NULL} set return_input to FALSE. Refer to \link[tidygeocoder]{api_parameter_reference} for more details.}
    \item{\code{return_addresses}}{if TRUE return input addresses. Defaults to TRUE if \code{return_input} is FALSE
and FALSE if \code{return_input} is TRUE. This argument is passed to the \code{geo()} function.}
    \item{\code{unique_only}}{if TRUE then only unique results will be returned and
return_input will be set to FALSE.}
  }}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{sf} object with POINT geometry for all geocoded addresses with valid coordinates.
}
\description{
Wraps \code{\link[tidygeocoder:geo]{tidygeocoder::geo()}} and \code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}} to convert a
character string or a data frame with an address column. Additional
parameters passed to \code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}} which passes \code{...} parameters to
\code{\link[tidygeocoder:geo]{tidygeocoder::geo()}}.
}
\seealso{
\code{\link[tidygeocoder:geo]{tidygeocoder::geo()}}, \code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}}
}
