% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_crs.R
\name{as_crs}
\alias{as_crs}
\alias{is_same_crs}
\alias{is_wgs84}
\title{Convert object to coordinate reference system or check coordinate reference
system}
\usage{
as_crs(x = NULL, allow_na = TRUE, arg = caller_arg(x), call = caller_env())

is_same_crs(x, y)

is_wgs84(x)
}
\arguments{
\item{x}{For \code{\link[=as_crs]{as_crs()}}, object to convert to a coordinate reference system.
For \code{\link[=is_same_crs]{is_same_crs()}} and \code{\link[=is_wgs84]{is_wgs84()}}, object to check. For \code{\link[=as_wgs84]{as_wgs84()}},
object to convert to EPSG:4326.}

\item{allow_na}{For \code{as_crs()}, if \code{TRUE}, return \code{NA_crs_} if x can't be
converted to a valid coordinate reference system. If \code{FALSE}, error instead
of returning an invalid CRS.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{y}{For \code{\link[=is_same_crs]{is_same_crs()}}, object to compare to x.}
}
\description{
Convert object to coordinate reference system or check coordinate reference
system
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))
nc[["category"]] <- sample(c("A", "B", "C"), nrow(nc), replace = TRUE)

as_sf(nc$geometry)

nc_bbox <- as_bbox(nc)

nc_bbox

as_sfc(nc_bbox)

as_xy(nc[1,])

as_sf_list(nc, col = "category")
}
