% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_bbox}
\alias{as_sfc}
\alias{as_sf_list}
\title{Convert an object to a simple feature or bounding box object}
\usage{
as_sf(x, crs = NULL, sf_col = "geometry", ...)

as_bbox(x, crs = NULL, ...)

as_sfc(x, crs = NULL, ...)

as_sf_list(x, nm = "data", col = NULL, crs = NULL, clean_names = TRUE)
}
\arguments{
\item{x}{A `sf`, `bbox`, `sfc`, `raster`, `sp`, or data frame object that can
be converted into a simple feature or bounding box object. [as_bbox()] can
also convert a vector with xmin, ymin, xmax, and ymax values.
[as_sf_list()] only supports sf objects or a data frames with a sf list
column named "data" (typically created by using [dplyr::group_nest()] on an
sf object.}

\item{crs}{Coordinate reference system for `sf`, `bbox`, `sfc` or `sf` list
object to return.}

\item{sf_col}{A column name to use for the geometry column created by
[as_sf]; defaults to "geometry".}

\item{...}{Additional parameters passed to [sf::st_bbox()] when calling
[as_bbox()] or passed to [sf::st_sf()], [sf::st_as_sf()], or [df_to_sf()]
for [as_sf()] (depending on class of x)}

\item{nm}{For [as_sf_list], name(s) for sf list; defaults to "data". If col
is provided, the values of the grouping column are used as names.}

\item{col}{For [as_sf_list], the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}

\item{clean_names}{If `TRUE`, clean names provided to nm or created based on
value of col using [janitor::clean_names]. If `FALSE`, use names as
provided.}
}
\description{
Both functions will pass a NULL value without returning an error. If a POINT
or MULTIPOINT object is passed to [as_bbox] a 0.00000001 meter buffer is
applied.
}
