% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_bbox}
\alias{as_sfc}
\alias{as_sf_class}
\title{Convert an object to a simple feature or bounding box object}
\usage{
as_sf(
  x,
  crs = NULL,
  sf_col = "geometry",
  ext = TRUE,
  ...,
  as_tibble = TRUE,
  call = caller_env()
)

as_bbox(x, crs = NULL, ext = TRUE, ..., call = caller_env())

as_sfc(x, crs = NULL, ext = TRUE, ..., call = caller_env())

as_sf_class(x, class = NULL, allow_null = TRUE, ..., call = caller_env())
}
\arguments{
\item{x}{A \code{sf}, \code{bbox}, \code{sfc}, \code{raster}, \code{sp}, or data frame object that can
be converted into a simple feature or bounding box object. \code{\link[=as_bbox]{as_bbox()}} can
also convert a vector with xmin, ymin, xmax, and ymax values.
\code{\link[=as_sf_list]{as_sf_list()}} only supports sf objects or a data frames with a sf list
column named "data" (typically created by using \code{\link[dplyr:group_nest]{dplyr::group_nest()}} on an
sf object.}

\item{crs}{Coordinate reference system for \code{sf}, \code{bbox}, \code{sfc} or \code{sf} list
object to return.}

\item{sf_col}{A column name to use for the geometry column created by
\link{as_sf}; defaults to "geometry".}

\item{ext}{If \code{TRUE}, \link{as_sf} will convert a data frame or character vector
of addresses to an sf object using \link{df_to_sf} or \link{address_to_sf}. If
\code{FALSE}, only spatial objects (bbox, sfg, sfc, sf list, raster, or sp
objects) can be converted. Defaults to \code{TRUE}.}

\item{...}{Additional parameters passed to \link{as_sf}, \link{as_sfc}, \link{as_bbox}, or
\link{as_sf_list}}

\item{as_tibble}{If \code{TRUE}, always return sf object as a tibble. If \code{FALSE},
some conversions may still return a tibble object.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{class}{A class to convert data to; defaults to NULL (which returns
"sf")}

\item{allow_null}{For \link{as_sf_class}, if class is \code{NULL} and allow_null is \code{TRUE},
return x without any class conversion or checks. Defaults to \code{TRUE}.}

\item{data}{Data that can be converted to sf, sfc, bbox or a sf list object.}
}
\description{
Both functions will pass a \code{NULL} value without returning an error. If a
POINT or MULTIPOINT object is passed to \code{\link[=as_bbox]{as_bbox()}} a 0.00000001 meter buffer
is applied. If a character object is passed to \code{\link[=as_bbox]{as_bbox()}} it is passed to
\code{\link[osmdata:getbb]{osmdata::getbb()}} using \code{format_out = "matrix"} which is converted into a
bounding box.
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))
nc[["category"]] <- sample(c("A", "B", "C"), nrow(nc), replace = TRUE)

as_sf(nc$geometry)

nc_bbox <- as_bbox(nc)

nc_bbox

as_sfc(nc_bbox)

as_xy(nc[1,])

as_sf_list(nc, col = "category")
}
