% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_bbox}
\alias{as_sfc}
\alias{as_sf_list}
\alias{as_sf_class}
\title{Convert an object to a simple feature or bounding box object}
\usage{
as_sf(x, crs = NULL, sf_col = "geometry", ext = TRUE, ...)

as_bbox(x, crs = NULL, ext = TRUE, ...)

as_sfc(x, crs = NULL, ext = TRUE, ...)

as_sf_list(x, nm = "data", col = NULL, crs = NULL, clean_names = TRUE)

as_sf_class(x, class = NULL, null.ok = TRUE, call = caller_env(), ...)
}
\arguments{
\item{x}{A \code{sf}, \code{bbox}, \code{sfc}, \code{raster}, \code{sp}, or data frame object that can
be converted into a simple feature or bounding box object. \code{\link[=as_bbox]{as_bbox()}} can
also convert a vector with xmin, ymin, xmax, and ymax values.
\code{\link[=as_sf_list]{as_sf_list()}} only supports sf objects or a data frames with a sf list
column named "data" (typically created by using \code{\link[dplyr:group_nest]{dplyr::group_nest()}} on an
sf object.}

\item{crs}{Coordinate reference system for \code{sf}, \code{bbox}, \code{sfc} or \code{sf} list
object to return.}

\item{sf_col}{A column name to use for the geometry column created by
\link{as_sf}; defaults to "geometry".}

\item{...}{Additional parameters passed to \link{as_sf}, \link{as_sfc}, \link{as_bbox}, or
\link{as_sf_list}}

\item{nm}{For \link{as_sf_list}, name(s) for sf list; defaults to "data". If col
is provided, the values of the grouping column are used as names.}

\item{col}{For \link{as_sf_list}, the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{class}{A class to convert data to; defaults to NULL (which returns
"sf")}

\item{data}{Data that can be converted to sf, sfc, bbox or a sf list object.}
}
\description{
Both functions will pass a \code{NULL} value without returning an error. If a POINT
or MULTIPOINT object is passed to \link{as_bbox} a 0.00000001 meter buffer is
applied.
}
