% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_xy}
\alias{as_xy}
\title{Convert data to a data frame with X/Y coordinate pairs}
\usage{
as_xy(x, data = NULL, crs = NULL, nm = c("x", "y"), ...)
}
\arguments{
\item{x}{A length 2 character string or numeric coordinate pair or a \code{sf},
\code{sfc}, or a \code{bbox} object. If x is a character string (e.g. c("xmin",
"ymax")), a sf, sfc, or bbox object must be provided to data argument.}

\item{nm}{Column names to use for X and Y columns.}
}
\description{
Wraps \code{\link[=as_points]{as_points()}}, \code{\link[=as_sfc]{as_sfc()}}, and \code{\link[=sf_bbox_point]{sf_bbox_point()}} to allow the
conversion of sf, sfc, or sfg objects into a simple data frame with X and Y
columns matching the provided nm parameter.
}
