% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dist.R
\name{compare_dist}
\alias{compare_dist}
\title{Does this distance fit within the bounding box of x or another distance?}
\usage{
compare_dist(
  x,
  dist = NULL,
  units = NULL,
  to = "xdist",
  check = TRUE,
  round = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{sf}, \code{sfc}, or \code{bbox} object or a numeric vector.}

\item{dist}{Distance to check as a units class or numeric vector}

\item{units}{Units to compare. If dist or x are numeric they are assumed to
be in the provided units or the units of the coordinate reference system
(CRS) for x (if x is a \code{sf}, \code{sfc}, or \code{bbox} object).}

\item{to}{Distance to compare dist with. Standard options include "xdist"
(default), "ydist", or "diagdist" to compare to \code{\link[=sf_bbox_xdist]{sf_bbox_xdist()}},
\code{\link[=sf_bbox_ydist]{sf_bbox_ydist()}}, or \code{\link[=sf_bbox_diagdist]{sf_bbox_diagdist()}}. If to is a length 2 character
vector, x, to, and any ... parameters are passed to \code{\link[=sf_bbox_dist]{sf_bbox_dist()}}. If to is
an sf or sfc object, to and any additional ... parameters are passed to
\code{\link[sf:geos_measures]{sf::st_distance()}} using x as the first argument and to as the second. If
x is a numeric vector the value of to is ignored.}

\item{check, round}{If check is \code{TRUE} (default), return a logical vector
indicating if the value dist can fit within the distance specified by x and
to. If check is \code{FALSE} and round is \code{TRUE}, return the number of times,
dist fits in the specified distance or 0 if dist is greater than the
specified distance. If check is \code{FALSE} and round is \code{FALSE}, return the
comparison distance divided by the input distance.}
}
\value{
A logical or numeric vector.
}
\description{
Check if a distance fits within the bounding box of a sf, sfc, or bbox
object, a distance returned by \code{\link[sf:geos_measures]{sf::st_distance()}} passing x as the first
parameter, or a units class or numeric value. Set \code{drop = FALSE} to return
the number of times dist fits within the specified distance.
}
\seealso{
Other dist: 
\code{\link{convert_dist_scale}()},
\code{\link{convert_dist_units}()},
\code{\link{get_measurements}},
\code{\link{is_dist_units}()},
\code{\link{sf_bbox_misc}}
}
\concept{dist}
