% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dist_units.R
\name{convert_dist_units}
\alias{convert_dist_units}
\title{Convert distance (and area) values between different units}
\usage{
convert_dist_units(dist, from = NULL, to = "meter")
}
\arguments{
\item{dist}{Numeric or units object}

\item{from}{Existing unit for dist, Default: `NULL`. If dist is a units object,
the numerator is used as "from"}

\item{to}{Unit to convert distance to, Default: 'meter'}
}
\value{
Object created by [units::set_units()]
}
\description{
Convert distance (and area) values between different units
}
\seealso{
Other dist: 
\code{\link{convert_dist_scale}()},
\code{\link{is_dist_units}()},
\code{\link{sf_bbox_misc}}
}
\concept{dist}
