% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_sf.R
\name{coords_to_sf}
\alias{coords_to_sf}
\title{Convert a data frame with one or more coordinate columns to an sf object}
\usage{
coords_to_sf(
  x,
  coords = c("lon", "lat"),
  into = NULL,
  sep = ",",
  rev = FALSE,
  remove_coords = FALSE,
  crs = 4326,
  call = caller_env()
)
}
\arguments{
\item{coords}{Coordinate columns for input data.frame or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{remove_coords}{For \code{\link[=df_to_sf]{df_to_sf()}}, if \code{TRUE}, remove the coordinate columns
after converting a data frame to simple feature object; defaults to
\code{FALSE}.}

\item{crs}{Coordinate reference system used by the coordinates in the
provided data frame.}

\item{keep_all}{If \code{FALSE}, drop all columns other than those named in
coords, Default: \code{TRUE}.}
}
\description{
\itemize{
\item \code{\link[=coords_to_sf]{coords_to_sf()}}: Convert a data frame with coordinates into a simple
feature object
\item \code{\link[=check_coords]{check_coords()}}: Check if a provided vector with coordinate column names
are valid for the provided data frame
\item \code{\link[=separate_coords]{separate_coords()}}: Separate coordinates from a single combined column
into two columns
\item \code{\link[=format_coords]{format_coords()}}: Format coordinates as numeric values and remove missing
coordinates from data frame
\item \code{\link[=has_coords]{has_coords()}}: Suggests a coordinate pair by comparing common values to
the column names for a provided data frame
\item \code{\link[=rev_coords]{rev_coords()}}: Reverse a vector with coordinate column names if latitude
or y appear first
}
}
