% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{count_features}
\alias{count_features}
\title{Count simple feature objects}
\usage{
count_features(
  x = NULL,
  y = NULL,
  join = NULL,
  by = NULL,
  .id = "count",
  col = NULL,
  ...
)
}
\arguments{
\item{join}{The join function used by \code{\link[sf:st_join]{sf::st_join()}} if data is an \code{sf}
object, Default: \link[sf:geos_binary_pred]{sf::st_intersects}}

\item{by}{Character string to join data by if data is not an \code{sf} object,
Default: \code{NULL}}

\item{.id}{Count column, Default: "count"}

\item{...}{Additional parameters (not in use)}

\item{data}{Data frame or \code{sf} object, Default: \code{NULL}}

\item{boundaries}{Boundary data as an \code{sf} object, Default: \code{NULL}}
}
\description{
Count the number of simple feature objects within a geometry or by some
attribute contained in both the data and the corresponding boundary geometry.
}
