% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asp.R
\name{get_asp}
\alias{get_asp}
\title{Get aspect ratio from string or based on specific paper and margins}
\usage{
get_asp(
  asp = NULL,
  paper = NULL,
  orientation = NULL,
  margin = NULL,
  unit = NULL,
  block_asp = FALSE,
  null.ok = TRUE
)
}
\arguments{
\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}
}
\value{
A numeric aspect ratio.
}
\description{
Get aspect ratio from string or based on specific paper and margins
}
