% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_margin.R
\name{get_margin}
\alias{get_margin}
\title{Get margins for a ggplot2 plot or map based on style or distance}
\usage{
get_margin(
  margin = NULL,
  paper = NULL,
  orientation = NULL,
  dist = NULL,
  unit = "in",
  block_width = NULL,
  header = 0,
  footer = 0
)
}
\arguments{
\item{margin}{Margin style (options include "extrawide", "wide", "standard",
"narrow", "none"), Additional "auto" option to generate margin based on
line length is planned but not yet implemented. Default: \code{NULL} (equivalent
to "none").}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}

\item{dist}{Margin distance (single value used to all sides), Default: \code{NULL}}

\item{unit}{Unit for margin distance, Default: 'in'.}

\item{block_width}{Plot or map width in units. If \code{paper} and \code{block_width} are
provided, margins are half the distance between the two evenly distributed.
This sets the margin distance for height as well as width so does not work
well with header and footers and should be improved in the future.}

\item{header, footer}{Header and footer height in units; defaults to 0. Please
note: headers and footers are not currently supported for "px" units.}
}
\value{
A \code{\link[ggplot2:element]{ggplot2::margin()}} element intended for use with
\code{\link[ggplot2:element]{ggplot2::element_rect()}} and the \code{plot.background} theme element.
}
\description{
This function works in combination with the \code{\link[=get_paper]{get_paper()}} function to make it
easier to position a map on a page before saving to file. This is primarily
useful when using a map or plot created with ggplot2 as part of a print
document format that is composed outside of R using a page layout application
such as Adobe InDesign.
}
\examples{
get_margin("standard")

get_margin("none")

get_margin(dist = 25, unit = "mm")

get_margin(paper = "letter", block_width = 5.5)
}
\seealso{
\code{\link[ggplot2:element]{ggplot2::margin()}}
}
