% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_geom_type.R
\name{is_geom_type}
\alias{is_geom_type}
\alias{is_point}
\alias{is_multipoint}
\alias{is_line}
\alias{is_multiline}
\alias{is_polygon}
\alias{is_multipolygon}
\title{What geometry type is this feature?}
\usage{
is_geom_type(x, type = NULL, by_geometry = FALSE, ext = TRUE)

is_point(x, by_geometry = FALSE)

is_multipoint(x, by_geometry = FALSE)

is_line(x, by_geometry = FALSE)

is_multiline(x, by_geometry = FALSE)

is_polygon(x, by_geometry = FALSE)

is_multipolygon(x, by_geometry = FALSE)
}
\arguments{
\item{x}{A \code{sf} or \code{sfc} object passed to \link[sf:st_geometry_type]{sf::st_geometry_type}}

\item{type}{If "POINT", check if geometry type is POINT. Same for all
available geometry types; not case sensitive; Default: \code{NULL}}

\item{by_geometry}{Passed to \link[sf:st_geometry_type]{sf::st_geometry_type}; defaults to \code{FALSE}}

\item{ext}{For st_geom_type, if ext \code{TRUE} and check is \code{NULL}, return a list
with checks for POINTS, POLYGONS, LINESTRING, and the returned types.}
}
\value{
If ext is FALSE and type is NULL, returns vector with geometry types
identical to \link[sf:st_geometry_type]{sf::st_geometry_type}. If ext is \code{TRUE}, returns a list and,
if type is not \code{NULL}, returns a logical vector.
}
\description{
A flexible wrapper for \link[sf:st_geometry_type]{sf::st_geometry_type} and \link[sf:st_is]{sf::st_is}.
}
