% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat_to_sfc.R
\name{lonlat_to_sfc}
\alias{lonlat_to_sfc}
\title{Convert a lon/lat or lat/lon coordinate pair to a sfc object}
\usage{
lonlat_to_sfc(
  x,
  range = getOption("sfext.coord_range", c(xmin = -180, ymin = -50, xmax = 180, ymax =
    60)),
  quiet = FALSE,
  call = parent.frame(),
  ...
)
}
\arguments{
\item{x}{A length 2 numeric vector with geodetic coordinates in a
\href{https://epsg.org/crs_4326/WGS-84.html}{EPSG:4326} coordinate reference
system.}

\item{range}{For \code{\link[=lonlat_to_sfc]{lonlat_to_sfc()}}, an object that is coercible to a \code{bbox}
object or a length 4 vector with names xmin, xmax, ymin, and ymax. If a
coordinate pair falls outside the latitude/longitude range defined by the
vector but inside the range if reversed, the coordinates are assumed to be
in lat/lon order and are switched to lon/lat order before being converted
to a point. Defaults to \code{c("xmin" = -180, "ymin" = -50, "xmax" = 180, "ymax" = 60)}. Note that this default setting will reverse valid
coordinates north of Anchorage, Alaska or south of New Zealand.}

\item{quiet}{If \code{TRUE}, suppress alert messages when converting a lat/lon
coordinate pair to a lon/lat pair. Defaults to \code{FALSE}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{...}{
  Arguments passed on to \code{\link[sf:sfc]{sf::st_sfc}}
  \describe{
    \item{\code{precision}}{numeric; see \link[sf]{st_as_binary}}
    \item{\code{check_ring_dir}}{see \link[sf]{st_read}}
    \item{\code{dim}}{character; if this function is called without valid geometries, this argument may carry the right dimension to set empty geometries}
    \item{\code{recompute_bbox}}{logical; use \code{TRUE} to force recomputation of the bounding box}
    \item{\code{oriented}}{logical; if \code{TRUE}, the ring is oriented such that left of the edges is inside the polygon; this is
needed for convering polygons larger than half the globe to s2}
    \item{\code{fall_back_class}}{character; class for return object when no geometries are provided as input}
  }}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=is_geo_coords]{is_geo_coords()}}
}
