% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{make_sf_grid_list}
\alias{make_sf_grid_list}
\title{Make a sf list by grid position}
\usage{
make_sf_grid_list(
  x,
  style = "rect",
  ncol = 2,
  nrow = 2,
  .id = "grid_id",
  crs = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{sf}, \code{sfc}, or \code{bbox} object, Default: \code{NULL}. Required.}

\item{style}{Style of cell to return with options including "rect", "square",
"hex", "flat_top_hex", "circle", "circle_offset"}

\item{ncol, nrow}{Used to set n if either are not \code{NULL}; defaults to \code{NULL}.
row and id are added as columns to the grid if they are provided.}

\item{.id}{A name to use for the cell id column. Defaults to "id".}

\item{crs}{Coordinate reference system of bounding box to return}

\item{...}{Additional parameters passed to \link{as_sf}, \link{as_sfc}, \link{as_bbox}, or
\link{as_sf_list}}
}
\description{
Create a grid with \code{\link[=st_make_grid_ext]{st_make_grid_ext()}} and
}
