% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_ext.R
\name{mapview_ext}
\alias{mapview_ext}
\alias{mapview_exif}
\alias{mapview_popup_img}
\title{Use mapview to interactively explore spatial data}
\usage{
mapview_ext(x, zcol = NULL, remove_na = FALSE, ...)

mapview_exif(
  path = NULL,
  fileext = "jpeg",
  popup = TRUE,
  tooltip = FALSE,
  images = NULL,
  width = 320,
  ...
)

mapview_popup_img(
  images,
  popup = TRUE,
  tooltip = FALSE,
  map = NULL,
  width = 320,
  ...,
  call = caller_env()
)
}
\arguments{
\item{x}{a \code{Raster*} or \code{Spatial*} or \code{Satellite} or
\code{sf} or \code{stars} object or a list of any combination of those. Furthermore,
this can also be a \code{data.frame}, a \code{numeric vector} or a
\code{character string} pointing to a tile image folder or file on disk.
If missing, a blank map will be drawn. A value of NULL will return NULL.}

\item{zcol}{attribute name(s) or column number(s) in attribute table
of the column(s) to be rendered. See also Details.}

\item{remove_na}{If \code{TRUE} and \code{zcol} is not \code{NULL}, filter \code{NA} values from
the data frame column defined by \code{zcol} before passing to
\code{\link[mapview:mapview]{mapview::mapview()}}}

\item{...}{
  Arguments passed on to \code{\link[mapview:mapview]{mapview::mapview}}
  \describe{
    \item{\code{maxpixels}}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}
    \item{\code{col.regions}}{color (palette) pixels.
See \code{\link[lattice]{levelplot}} for details.}
    \item{\code{at}}{the breakpoints used for the visualisation.
See \code{\link[lattice]{levelplot}} for details.}
    \item{\code{na.color}}{color for missing values}
    \item{\code{use.layer.names}}{should layer names of the Raster* object be used?}
    \item{\code{map.types}}{character specifications for the base maps.
see \url{https://leaflet-extras.github.io/leaflet-providers/preview/}
for available options.}
    \item{\code{alpha.regions}}{opacity of the fills of points, polygons or raster layer(s)}
    \item{\code{legend}}{should a legend be plotted}
    \item{\code{legend.opacity}}{opacity of the legend}
    \item{\code{trim}}{should the raster be trimmed in case there are NAs on the edges}
    \item{\code{verbose}}{should some details be printed during the process}
    \item{\code{layer.name}}{the name of the layer to be shown on the map.
By default this is the character version of whatever is passed to \code{x}.
NOTE: This is being passed to underlying leaflet functions as the group argument.
So if you use mapview to set up a map and want to refer to a certain layer
later on, this is what you should refer to in \code{group}.}
    \item{\code{homebutton}}{logical, whether to add a zoom-to-layer button to the map.
Defaults to TRUE}
    \item{\code{native.crs}}{logical whether to reproject to web map coordinate
reference system (web mercator - epsg:3857) or render using native CRS of
the supplied data (can also be NA). Default is FALSE which will render in
web mercator. If set to TRUE now background maps will be drawn (but rendering
may be much quicker as no reprojecting is necessary). Currently only works
for simple features.}
    \item{\code{method}}{for raster data only (raster/stars). Method used to compute
values for the resampled layer that is passed on to leaflet. mapview does
projection on-the-fly to ensure correct display and therefore needs to know
how to do this projection. The default is 'bilinear' (bilinear interpolation),
which is appropriate for continuous variables. The other option, 'ngb'
(nearest neighbor), is useful for categorical variables. Ignored if the raster
layer is of class \code{factor} in which case "ngb" is used.}
    \item{\code{label}}{For vector data (sf/sp) a character vector of labels to be
shown on mouseover. See \code{\link[leaflet]{addControl}} for details. For raster
data (Raster*/stars) a logical indicating whether to add image query.}
    \item{\code{query.type}}{for raster methods only. Whether to show raster value query
on \code{'mousemove'} or \code{'click'}. Ignored if \code{label = FALSE}.}
    \item{\code{query.digits}}{for raster methods only. The amount of digits to be shown
by raster value query. Ignored if \code{label = FALSE}.}
    \item{\code{query.position}}{for raster methods only. The position of the raster
value query info box. See \code{position} argument of \code{\link[leaflet]{addLegend}}
for possible values. Ignored if \code{label = FALSE}.}
    \item{\code{query.prefix}}{for raster methods only. a character string to be shown
as prefix for the layerId. Ignored if \code{label = FALSE}.}
    \item{\code{viewer.suppress}}{deprecated.
Use \code{mapviewOptions(viewer.suppress = TRUE/FALSE)} instead.}
    \item{\code{hide}}{either a logical, a vector of layer names or a vector of layer indices.
See Details for more information on what exactly it does for different raster types.}
    \item{\code{band}}{for stars layers, the band number to be plotted.}
    \item{\code{pane}}{name of the map pane in which to render features. See
\code{\link[leaflet]{addMapPane}} for details. Currently only supported for vector layers.
Ignored if \code{canvas = TRUE}. The default \code{"auto"} will create different panes
for points, lines and polygons such that points overlay lines overlay polygons.
Set to \code{NULL} to get default leaflet behaviour where allfeatures
are rendered in the same pane and layer order is determined automatically/sequentially.}
    \item{\code{canvas}}{whether to use canvas rendering rather than svg. May help
performance with larger data. See \url{https://leafletjs.com/index.html#canvas}
for more information. Only applicable for vector data. The default setting will
decide automatically, based on feature complexity.}
    \item{\code{burst}}{whether to show all (TRUE) or only one (FALSE) layer(s).
See also Details.}
    \item{\code{color}}{color (palette) for points/polygons/lines}
    \item{\code{cex}}{attribute name(s) or column number(s) in attribute table
of the column(s) to be used for defining the size of circles}
    \item{\code{lwd}}{line width}
    \item{\code{alpha}}{opacity of lines}
    \item{\code{na.alpha}}{opacity of missing values}
    \item{\code{highlight}}{either \code{FALSE}, \code{NULL} or a list of styling
options for feature highlighting on mouse hover.
See \code{\link[leaflet]{highlightOptions}} for details.}
    \item{\code{maxpoints}}{the maximum number of points making up the geometry.
In case of lines and polygons this refers to the number of vertices. See
Details for more information.}
    \item{\code{tms}}{whether the tiles are served as TMS tiles.}
    \item{\code{y}}{numeric vector.}
    \item{\code{type}}{whether to render the numeric vector \code{x} as a
point \code{"p"} or line \code{"l"} plot.}
    \item{\code{grid}}{whether to plot a (scatter plot) xy-grid to aid interpretation
of the visualisation. Only relevant for the data.frame method.}
    \item{\code{xcol}}{the column to be mapped to the x-axis. Only relevant for the
data.frame method.}
    \item{\code{ycol}}{the column to be mapped to the y-axis. Only relevant for the
data.frame method.}
    \item{\code{aspect}}{the ratio of x/y axis coordinates to adjust the plotting
space to fit the screen. Only relevant for the data.frame method.}
    \item{\code{crs}}{an optional crs specification for the provided data to enable
rendering on a basemap. See argument description in \code{\link[sf]{st_sf}}
for details.}
  }}

\item{path}{A path to folder or file.}

\item{fileext}{File extension. Defaults to "jpeg".}

\item{popup}{If \code{TRUE}, add a popup image to a leaflet map; defaults \code{TRUE}.}

\item{tooltip}{logical, whether to show image(s) as popup(s) (on click) or
tooltip(s) (on hover).}

\item{images}{A simple feature object with columns for the image path/url,
image width, and image height.}

\item{width}{the width of the image(s) in pixels.}

\item{map}{an optional existing map to be updated/added to.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
A wrapper for \code{\link[mapview:mapview]{mapview::mapview()}} that drops list columns and makes it
easier to quickly specify a zcol value.
}
\seealso{
\code{\link[mapview:mapview]{mapview::mapview()}}
}
