% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_ext.R
\name{mapview_ext}
\alias{mapview_ext}
\alias{mapview_exif}
\alias{mapview_popup_img}
\title{Use mapview to interactively explore spatial data}
\usage{
mapview_ext(x, zcol = NULL, remove_na = FALSE, ...)

mapview_exif(
  path = NULL,
  fileext = "jpeg",
  popup = TRUE,
  tooltip = FALSE,
  images = NULL,
  width = 320,
  ...
)

mapview_popup_img(
  images,
  popup = TRUE,
  tooltip = FALSE,
  map = NULL,
  width = 320,
  ...,
  call = caller_env()
)
}
\arguments{
\item{x}{a \code{Raster*} or \code{Spatial*} or \code{Satellite} or
\code{sf} or \code{stars} object or a list of any combination of those. Furthermore,
  this can also be a \code{data.frame}, a \code{numeric vector} or a
\code{character string} pointing to a tile image folder or file on disk.
  If missing, a blank map will be drawn. A value of NULL will return NULL.}

\item{zcol}{attribute name(s) or column number(s) in attribute table
of the column(s) to be rendered. See also Details.}

\item{remove_na}{If TRUE and zcol is not \code{NULL}, filter \code{NA} values from the
zcol before passing to \code{\link[mapview:mapView]{mapview::mapview()}}}

\item{...}{
  Arguments passed on to \code{\link[mapview:mapView]{mapview::mapview}}
  \describe{
    \item{\code{}}{}
  }}

\item{path}{A path to folder or file.}

\item{popup}{If \code{TRUE}, add a popup image to a leaflet map; defaults \code{TRUE}.}

\item{images}{A simple feature object with columns for the image path/url,
image width, and image height.}

\item{map}{an optional existing map to be updated/added to.}
}
\description{
A wrapper for \code{\link[mapview:mapView]{mapview::mapview()}} that drops list columns and makes it
easier to quickly specify a zcol value.
}
\seealso{
\code{\link[mapview:mapView]{mapview::mapview()}}
}
