% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{misc_sf}
\alias{misc_sf}
\alias{transform_sf}
\alias{relocate_sf_col}
\alias{rename_sf_col}
\alias{get_sf_col}
\title{Additional utility functions for sf objects}
\usage{
transform_sf(x, crs = NULL, null.ok = TRUE, ...)

relocate_sf_col(x, .after = dplyr::everything())

rename_sf_col(x, sf_col = "geometry")

get_sf_col(x = NULL)
}
\arguments{
\item{x}{A \code{sf} or \code{sfc} object. If x has a missing crs, the crs is set to
the provided value.}

\item{crs}{A coordinate reference system identifier (numeric or character) or
a \code{sf}, \code{sfc}, \code{bbox}, or \code{crs} class object supported by \code{\link[sf:st_crs]{sf::st_crs()}}.}

\item{null.ok}{If \code{TRUE} and crs is \code{NULL}, return x.}

\item{...}{Additional parameters passed to \code{\link[sf:st_transform]{sf::st_transform()}}}

\item{.after}{The location to place sf column after; defaults to
\link[dplyr:reexports]{dplyr::everything}.}

\item{sf_col}{Name to use for the sf column after renaming; defaults to "geometry".}
}
\description{
Additional utility functions for sf objects
}
