% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{misc_sf}
\alias{misc_sf}
\alias{transform_sf}
\alias{relocate_sf_col}
\alias{rename_sf_col}
\title{Additional utility functions for sf objects}
\usage{
transform_sf(x, crs = NULL, null.ok = TRUE, ...)

relocate_sf_col(x, .after = dplyr::everything())

rename_sf_col(x, sf_col = "geometry")
}
\arguments{
\item{x}{A \code{sf} object.}

\item{crs}{A coordinate reference system or an sf, sfc, or bbox object
returns a crs from \link[sf:st_crs]{sf::st_crs}.}

\item{null.ok}{If \code{TRUE} and x is \code{NULL}, return x.}

\item{...}{Additional parameters passed to \link[sf:st_transform]{sf::st_transform}}

\item{.after}{The location to place sf column after; defaults to
\link[dplyr:reexports]{dplyr::everything}.}

\item{sf_col}{Name to use for the sf column after renaming; defaults to "geometry".}
}
\description{
Additional utility functions for sf objects
}
