% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_exif.R
\name{read_sf_exif}
\alias{read_sf_exif}
\title{Read EXIF metadata to create a simple feature object or write
EXIF metadata to image files}
\usage{
read_sf_exif(
  path = NULL,
  fileext = NULL,
  filetype = NULL,
  bbox = NULL,
  sort = NULL,
  tags = NULL,
  geometry = TRUE,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{path}{A path to folder or file.}

\item{fileext, filetype}{File extension or file type. filetype is used if
fileext is \code{NULL}.}

\item{bbox}{Bounding box to filter by.}

\item{sort}{Column name for variable to sort by passed to \code{\link[=sort_features]{sort_features()}}.
Currently supports "lon", "lat", or "filename". Defaults to \code{NULL}.}

\item{tags}{Optional list of EXIF tags to read from files. Must include GPS
tags to create an \code{sf} object.}

\item{geometry}{If \code{TRUE} (default), return a simple feature object. If
\code{FALSE}, return a data.frame.}

\item{quiet}{If \code{TRUE} (default), suppress function messages.}

\item{...}{Additional parameters to pass to \code{\link[exiftoolr:exif_read]{exiftoolr::exif_read()}}}
}
\description{
\code{read_sf_exif()} read EXIF data from folder of files and, geometry is \code{TRUE}
and coordinate metadata is available, convert the data to a sf object. This
function also assigns a cardinal direction based on the direction metadata
and recodes the orientation metadata.

For \code{write_exif()} the parameters are used to multiple tags with the same
values:
\itemize{
\item title: Title, IPTC:Headline, IPTC:ObjectName, XMP-dc:Title
\item description: ImageDescription, XMP-dc:Description, and
IPTC:Caption-Abstract
\item keywords: Keywords, IPTC:Keywords, XMP-dc:Subject
}
}
\examples{
\dontrun{
  read_sf_exif(
    path = system.file("extdata/photos", package = "overedge"),
    filetype = "jpeg"
  )
}
}
\seealso{
Other read_write: 
\code{\link{read_sf_ext}()}
}
\concept{read_write}
