% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_exif.R
\name{read_sf_exif}
\alias{read_sf_exif}
\alias{write_exif}
\alias{write_exif_keywords}
\title{Read location data from images to a simple feature object using EXIF or write EXIF metadata}
\usage{
read_sf_exif(
  path = NULL,
  filetype = NULL,
  bbox = NULL,
  sort = "lon",
  tags = NULL,
  ...
)

write_exif(
  path = NULL,
  filetype = NULL,
  title = NULL,
  author = NULL,
  date = NULL,
  keywords = NULL,
  args = NULL,
  overwrite = TRUE
)

write_exif_keywords(
  path,
  filetype = NULL,
  key_list,
  key_col = "name",
  keywords = NULL,
  join = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{path}{A path to folder of one or more files with EXIF location
metadata.}

\item{filetype}{The file extension or file type; defaults to `NULL`.}

\item{bbox}{Optional bounding box to crop returned file (excluding images
with location data outside the bounding box). If bbox is provided the
returned data will match the crs of the bbox.}

\item{sort}{Column name for variable to sort by. Currently supports "lon"
(default), "lat", or "filename"}

\item{tags}{Optional list of EXIF tags to read from files. Must include GPS
tags to create an sf object.}

\item{...}{Additional EXIF tags to pass to [exiftoolr::exif_read]}

\item{title}{Title to add to file metadata with exiftoolr, Default: `NULL`.}

\item{author}{Author to add to file metadata with exiftoolr, Default: `NULL`.}

\item{date}{Date to add to file metadata with exiftoolr (not currently
working), Default: `NULL`.}

\item{keywords}{Keyword(s) added to file metadata with with exiftoolr,
Default: `NULL`.}

\item{args}{Alternate arguments passed to [exiftoolr::exif_call()]. If args
is not `NULL`, title, author, date, and keywords are ignored; defaults to
`NULL`.}

\item{overwrite}{If TRUE, overwrite any existing EXIF metadata present in the
provided fields; defaults to TRUE}

\item{key_list}{List of sf objects with features with keywords, e.g. boundaries}

\item{key_col}{Column name in key_list with the values to use for keywords.}

\item{join}{geometry predicate function; defaults to `NULL`, set to
[sf::st_intersects] if key_list contains only POLYGON or MULTIPOLYGON objects
or [sf::st_nearest_feature] if key_list contains other types.}
}
\description{
Read EXIF data from folder of images.
}
\section{Writing EXIF metadata}{
NA
}

\examples{
read_sf_exif(
  path = system.file("extdata/photos", package = "overedge"),
  filetype = "jpeg"
)
}
\seealso{
Other read_write: 
\code{\link{read_sf_ext}()}
}
\concept{read_write}
