% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ext.R
\name{read_sf_ext}
\alias{read_sf_ext}
\alias{read_sf_pkg}
\alias{read_sf_path}
\alias{read_sf_query}
\alias{read_sf_excel}
\alias{read_sf_csv}
\alias{read_sf_url}
\alias{read_sf_esri}
\alias{read_sf_gist}
\alias{read_sf_gmap}
\alias{read_sf_download}
\alias{read_sf_gsheet}
\title{Read spatial data in a bounding box to a simple feature object from multiple sources}
\usage{
read_sf_ext(...)

read_sf_pkg(data, bbox = NULL, package = NULL, filetype = "gpkg", ...)

read_sf_path(path, bbox = NULL, ...)

read_sf_query(
  path,
  dsn = NULL,
  bbox = NULL,
  query = NULL,
  table = NULL,
  name = NULL,
  name_col = NULL,
  wkt_filter = NULL,
  zm_drop = FALSE,
  ...
)

read_sf_excel(
  path,
  sheet = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  ...
)

read_sf_csv(
  path,
  url = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  show_col_types = FALSE,
  ...
)

read_sf_url(url, bbox = NULL, coords = NULL, zm_drop = TRUE, ...)

read_sf_esri(
  url,
  bbox = NULL,
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  ...
)

read_sf_gist(url, id = NULL, bbox = NULL, ...)

read_sf_gmap(url, bbox = NULL, zm_drop = TRUE)

read_sf_download(
  url,
  filename,
  bbox = NULL,
  path = NULL,
  filetype = "geojson",
  prefix = "date",
  method = "auto",
  unzip = FALSE,
  ...
)

read_sf_gsheet(
  url,
  sheet = NULL,
  ss = NULL,
  bbox = NULL,
  ask = FALSE,
  coords = c("lon", "lat"),
  from_crs = 4326,
  geo = FALSE,
  address = "address",
  ...
)
}
\arguments{
\item{...}{Additional parameters passed to multiple functions; see details.}

\item{data}{Name of a package dataset; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only.}

\item{bbox}{A bounding box object; defaults to \code{NULL}. If \code{"bbox"} is provided,
only returns features intersecting the bounding box.}

\item{package}{Package name; used by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only.}

\item{filetype}{File type supported by \code{\link[sf:st_read]{sf::read_sf()}}; Default: 'gpkg'; used
by \code{\link[=read_sf_pkg]{read_sf_pkg()}} only and required only if the data is in the package
cache directory or extdata system files.}

\item{path}{A file path.}

\item{dsn}{data source name (interpretation varies by driver - for some
drivers, \code{dsn} is a file name, but may also be a folder, or contain
the name and access credentials of a database); in case of GeoJSON,
\code{dsn} may be the character string holding the geojson data. It can
also be an open database connection.}

\item{query}{SQL query to select records; see details}

\item{name}{Name to make file name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}, e.g. "Residential zoning map" becomes
"residential_zoning_map"}

\item{wkt_filter}{character; WKT representation of a spatial filter (may be used as bounding box, selecting overlapping geometries); see examples}

\item{zm_drop}{If \code{TRUE}, drop Z and/or M dimensions using \link[sf:st_zm]{sf::st_zm}}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{coords}{Character vector with coordinate values. Coordinates must use
the same crs as the \code{from_crs} parameter.}

\item{from_crs}{For \link{df_to_sf}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{geo}{If \code{TRUE}, use \link{address_to_sf} to geocode address column; defaults
to \code{FALSE}.}

\item{address}{Address column name, Default: 'address'}

\item{url}{A url for a spatial data file, tabular data with coordinates, or a
ArcGIS FeatureServer or MapServer to access with \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf()}}}

\item{show_col_types}{If \code{FALSE}, do not show the guessed column types. If
\code{TRUE} always show the column types, even if they are supplied. If \code{NULL}
(the default) only show the column types if they are not explicitly supplied
by the \code{col_types} argument.}

\item{id}{The name of a column in which to store the file path. This is
useful when reading multiple input files and there is data in the file
paths, such as the data collection date. If \code{NULL} (the default) no extra
column is created.}

\item{filename}{File name; if file name is \code{NULL}, name and file type are
both required.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{method}{the geocoding service to be used. API keys are loaded from environmental variables. Run \code{usethis::edit_r_environ()} to open your .Renviron file and add an API key as an environmental variable. For example, add the line \code{GEOCODIO_API_KEY="YourAPIKeyHere"}.
\itemize{
\item \code{"osm"}: \href{https://nominatim.org}{Nominatim}.
\item \code{"census"}: \href{https://geocoding.geo.census.gov/}{US Census}. Geographic coverage is limited to the United States.  Batch geocoding is supported.
\item \code{"arcgis"}: \href{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}{ArcGIS}.
\item \code{"geocodio"}: \href{https://www.geocod.io/}{Geocodio}. Geographic coverage is limited to the United States and Canada. An API key must be stored in the environmental variable "GEOCODIO_API_KEY". Batch geocoding is supported.
\item \code{"iq"}: \href{https://locationiq.com/}{Location IQ}.  An API key must be stored in the environmental variable "LOCATIONIQ_API_KEY".
\item \code{"google"}: \href{https://developers.google.com/maps/documentation/geocoding/overview}{Google}.  An API key must be stored in the environmental variable "GOOGLEGEOCODE_API_KEY".
\item \code{"opencage"}: \href{https://opencagedata.com}{OpenCage}.  An API key must be stored in the environmental variable "OPENCAGE_KEY".
\item \code{"mapbox"}: \href{https://docs.mapbox.com/api/search/}{Mapbox}.  An API key must be stored in the environmental variable "MAPBOX_API_KEY".
\item \code{"here"}: \href{https://developer.here.com/products/geocoding-and-search}{HERE}.  An API key must be stored in the environmental variable "HERE_API_KEY". Batch geocoding is supported, but must be explicitly called with \code{mode = "batch"}.
\item \code{"tomtom"}: \href{https://developer.tomtom.com/search-api/search-api-documentation/geocoding}{TomTom}.  An API key must be stored in the environmental variable "TOMTOM_API_KEY". Batch geocoding is supported.
\item \code{"mapquest"}: \href{https://developer.mapquest.com/documentation/geocoding-api/}{MapQuest}.  An API key must be stored in the environmental variable "MAPQUEST_API_KEY". Batch geocoding is supported.
\item \code{"bing"}: \href{https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/}{Bing}.  An API key must be stored in the environmental variable "BINGMAPS_API_KEY". Batch geocoding is supported, but must be explicitly called with \code{mode = "batch"}.
\item \code{"geoapify"}: \href{https://www.geoapify.com/geocoding-api}{Geoapify}.  An API key must be stored in the environmental variable "GEOAPIFY_KEY".
\item \code{"cascade"} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use \link[tidygeocoder]{geocode_combine} or \link[tidygeocoder]{geo_combine} instead.
The "cascade" method first uses one geocoding service and then uses
a second geocoding service if the first service didn't return results.
The services and order is specified by the cascade_order argument.
Note that this is not compatible with \code{full_results = TRUE} as geocoding
services have different columns that they return.
}}

\item{unzip}{If \code{TRUE}, url must be a zip file that is downloaded, unzipped
into a temporary directory (created with \code{\link[=tempdir]{tempdir()}}), and then read to a file using the specified
file type.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:as_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{ask}{If \code{TRUE}, ask for the name of the Google Sheet to read if ss is
not provided to \link{read_sf_gsheet}.}
}
\description{
An extended version of \code{\link[sf:st_read]{sf::read_sf()}} that support reading spatial data
based on a file path, URL, or the data name and associated package.
Optionally provide a bounding box to filter data (not supported for all data
types).
}
\details{
Reading data from a url:

\link{read_sf_url} supports multiple types of urls:
\itemize{
\item A MapServer or FeatureServer URL
\item A URL for a GitHub gist with a single spatial data file (first file used if gist contains multiple)
\item A URL for a spatial data file or a CSV file
\item A Google Sheets URL
\item A public Google Maps URL
}

Reading data from a package:

\link{read_sf_pkg} looks for three types of package data:
\itemize{
\item Data loaded with the package
\item External data in the \code{extdata} system files folder.
\item Cached data in the cache directory returned by \link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir}
}

Additional ... parameters:

\link{read_sf_pkg} and \link{read_sf_download} both pass additional parameters
to \link{read_sf_path} which supports query, name_col, name, and table. name and
name_col are ignored if a query parameter is provided. If table is not
provided, a expected layer name is created based on the file path.

\link{read_sf_url} pass the where, name_col, and name for any ArcGIS FeatureServer or
MapServer url (passed to \link{read_sf_esri}) or sheet if the url is for a Google
Sheet (passed to \link[googlesheets4:range_read]{googlesheets4::read_sheet}), or a query or wkt filter
parameter if the url is some other type (passed to \link[sf:st_read]{sf::read_sf}).

\link{read_sf_ext} is a flexible function that only has bbox as a named parameter
and all other parameters in ... are passed to one of the other overedge
read_sf functions.
}
\seealso{
Other read_write: 
\code{\link{make_filename}()},
\code{\link{read_sf_exif}()}
}
\concept{read_write}
