% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox_misc.R
\name{sf_bbox_misc}
\alias{sf_bbox_misc}
\alias{sf_bbox_to_sf}
\alias{sf_bbox_to_sfc}
\alias{sf_bbox_transform}
\alias{sf_bbox_point}
\alias{sf_bbox_to_wkt}
\alias{sf_bbox_to_lonlat_query}
\alias{sf_bbox_to_npc}
\title{Convert and tranform bounding boxes}
\usage{
sf_bbox_to_sf(bbox, sf_col = "geometry")

sf_bbox_to_sfc(bbox)

sf_bbox_transform(bbox, crs = NULL)

sf_bbox_point(bbox, point = NULL, crs = NULL, call = caller_env())

sf_bbox_to_wkt(bbox, crs = NULL)

sf_bbox_to_lonlat_query(bbox, coords = c("longitude", "latitude"), crs = 4326)

sf_bbox_to_npc(bbox, point)
}
\arguments{
\item{bbox}{A bbox object.}

\item{sf_col}{name to use for geometry column after converting to simple
feature object; defaults to "geometry".}

\item{crs}{A coordinate reference system to use for
\code{\link[=sf_bbox_to_lonlat_query]{sf_bbox_to_lonlat_query()}} (defaults to 4326) or for resulting bounding
box (\code{\link[=sf_bbox_transform]{sf_bbox_transform()}}) or sfc object (\code{\link[=sf_bbox_point]{sf_bbox_point()}}) (defaults to
\code{NULL}).}

\item{point}{point to find npc coords for center}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{coords}{Column names with coordinates for query. e.g. \code{c("X", "Y")} or
\code{c("longitude", "latitude")} (default). Used by \code{\link[=sf_bbox_to_lonlat_query]{sf_bbox_to_lonlat_query()}}
only.}
}
\value{
sf object
}
\description{
Simple bounding box functions that you can use to:
\itemize{
\item Convert a bounding box to a simple feature object
(\code{\link[=sf_bbox_to_sf]{sf_bbox_to_sf()}}) or simple feature collection (\code{\link[=sf_bbox_to_sfc]{sf_bbox_to_sfc()}})
\item Transform the coordinate reference system (\code{\link[=sf_bbox_transform]{sf_bbox_transform()}})
\item Return a point from any of the corners, center, or midpoints
(\code{\link[=sf_bbox_point]{sf_bbox_point()}})
\item Convert a bounding box to a SQL style query (\code{\link[=sf_bbox_to_lonlat_query]{sf_bbox_to_lonlat_query()}}),
well known text (\code{\link[=sf_bbox_to_wkt]{sf_bbox_to_wkt()}})
\item Convert a point and a corresponding bounding box into into a npc
(normalised parent coordinates) value with \code{\link[=sf_bbox_to_npc]{sf_bbox_to_npc()}}
}

If sf (>=1.0-17) is installed, \code{\link[=sf_bbox_transform]{sf_bbox_transform()}} uses
\code{\link[sf:st_transform]{sf::st_transform()}} \emph{without} converting to a sfc object and back to bbox.
}
\seealso{
\itemize{
\item \code{\link[sf:st_as_sf]{sf::st_as_sf()}},\code{\link[sf:st_as_sfc]{sf::st_as_sfc()}}
\item \code{\link[sfx:st_extent]{sfx::st_extent()}}
}
}
