% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox_misc.R
\name{sf_bbox_misc}
\alias{sf_bbox_misc}
\alias{sf_bbox_asp}
\alias{sf_bbox_point}
\alias{sf_bbox_dist}
\alias{sf_bbox_xdist}
\alias{sf_bbox_ydist}
\alias{sf_bbox_diagdist}
\alias{sf_bbox_transform}
\alias{sf_bbox_to_sf}
\alias{sf_bbox_to_sfc}
\alias{sf_bbox_to_wkt}
\alias{sf_bbox_to_lonlat_query}
\alias{sf_bbox_shift}
\alias{sf_bbox_contract}
\alias{sf_bbox_expand}
\title{Measure, transform, and convert bounding boxes}
\usage{
sf_bbox_asp(bbox, orientation = FALSE)

sf_bbox_point(bbox, point = NULL, call = caller_env())

sf_bbox_dist(
  bbox,
  from,
  to,
  units = NULL,
  drop = TRUE,
  by_element = TRUE,
  call = caller_env(),
  ...
)

sf_bbox_xdist(bbox, units = NULL, drop = TRUE)

sf_bbox_ydist(bbox, units = NULL, drop = TRUE)

sf_bbox_diagdist(bbox, units = NULL, drop = TRUE)

sf_bbox_transform(bbox, crs = NULL)

sf_bbox_to_sf(bbox, sf_col = "geometry")

sf_bbox_to_sfc(bbox)

sf_bbox_to_wkt(bbox)

sf_bbox_to_lonlat_query(bbox, coords = c("longitude", "latitude"), crs = 4326)

sf_bbox_shift(
  bbox,
  x_nudge = 0,
  y_nudge = 0,
  side = c("all", "top", "bottom", "left", "right"),
  dir = NULL,
  call = caller_env()
)

sf_bbox_contract(bbox, x_nudge = 0, y_nudge = 0)

sf_bbox_expand(bbox, x_nudge = 0, y_nudge = 0)
}
\arguments{
\item{bbox}{A bbox object.}

\item{orientation}{If \code{TRUE}, sf_bbox_asp returns a suggested orientation
based on aspect ratio (< 0.9 "portrait"; > 1.1 "landscape"; else "square");
defaults to \code{FALSE}.}

\item{point}{Length 2 character vector representing a coordinate pair at a
corner, midpoint, or center of the bounding box ("xmin", "ymin", "xmax",
"ymax", "xmid", "ymid")}

\item{call}{Passed as the error_call parameter for \link[rlang:arg_match]{rlang::arg_match} to
improve error messages when function is used internally.}

\item{from, to}{xy pairs (e.g. c("xmax", "ymax) defining points to measure
distance from and to.}

\item{units}{The units to return for sf_bbox_dist. Defaults to NULL.}

\item{drop}{If \code{FALSE}, distance functions return with units. If \code{FALSE}
(default), distance functions return numeric values.}

\item{crs}{coordinate reference system to use for query; default 4326}

\item{sf_col}{name to use for geometry column after converting to simple
feature object; defaults to "geometry".}

\item{coords}{query column names with coordinates. e,g, c("X", "Y") or
c("lat", "lon")}

\item{x_nudge, y_nudge}{Length 1 or 2 numeric vector; unitless.}

\item{side}{one or more sides to shift: "top", "bottom", "left", "right", or
"all"}

\item{dir}{If "in", contract the \code{bbox} by x_nudge and y_nudge. If "out",
expand the bbox by x_nudge and y_nudge. If dir is not \code{NULL}; absolute
values are used for x_nudge and y_nudge. Defaults to \code{NULL}.}
}
\value{
sf object
}
\description{
Simple bounding box functions that you can use to:
}
\details{
\itemize{
\item Measure distances (\link{sf_bbox_dist}, \link{sf_bbox_xdist}, \link{sf_bbox_ydist}, and \link{sf_bbox_diagdist})
\item Get an aspect ratio or orientation (\link{sf_bbox_asp}) (counts asp between 0.9 and 1.1 as "square")
\item Return a point from any of the corners, center, or midpoints (\link{sf_bbox_point})
\item Transform the coordinate reference system (\link{sf_bbox_transform})
\item Convert a bounding box to a SQL style query (\link{sf_bbox_to_lonlat_query}), well
known text (\link{sf_bbox_to_wkt}), or a simple feature object (\link{sf_bbox_to_sf})
\item Shift, expand, or contract a bounding box (\link{sf_bbox_shift}, \link{sf_bbox_expand}, \link{sf_bbox_contract})
}

The only functions with additional parameters are sf_bbox_to_lonlat_query and
sf_bbox_dist. All other functions only take a bbox parameter.
}
\seealso{
\itemize{
\item \code{\link[sf:st_as_sf]{sf::st_as_sf()}},\code{\link[sf:st_as_sfc]{sf::st_as_sfc()}}
\item \code{\link[sfx:st_extent]{sfx::st_extent()}}
}

Other dist: 
\code{\link{compare_dist}()},
\code{\link{convert_dist_scale}()},
\code{\link{convert_dist_units}()},
\code{\link{get_measurements}},
\code{\link{is_dist_units}()}
}
\concept{dist}
