% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bbox_shift.R
\name{sf_bbox_shift}
\alias{sf_bbox_shift}
\alias{sf_bbox_contract}
\alias{sf_bbox_expand}
\title{Shift sides, contract, or expand a bounding box}
\usage{
sf_bbox_shift(
  bbox,
  x_nudge = 0,
  y_nudge = 0,
  side = c("all", "top", "bottom", "left", "right"),
  dir = NULL,
  call = caller_env()
)

sf_bbox_contract(bbox, x_nudge = 0, y_nudge = 0)

sf_bbox_expand(bbox, x_nudge = 0, y_nudge = 0)
}
\arguments{
\item{bbox}{A \code{bbox} object.}

\item{x_nudge, y_nudge}{Length 1 or 2 numeric vector; unitless. Used by or
passed to \code{\link[=sf_bbox_shift]{sf_bbox_shift()}}. Required for  \code{\link[=sf_bbox_shift]{sf_bbox_shift()}}.}

\item{side}{one or more sides to shift: "top", "bottom", "left", "right", or
"all". Required for \code{\link[=sf_bbox_shift]{sf_bbox_shift()}}.}

\item{dir}{If "in", contract the \code{bbox} by x_nudge and y_nudge. If "out",
expand the bbox by x_nudge and y_nudge. If dir is not \code{NULL}; absolute
values are used for x_nudge and y_nudge. Defaults to \code{NULL}. Optional
\code{\link[=sf_bbox_shift]{sf_bbox_shift()}}.}

\item{call}{Passed as the error_call parameter for \link[rlang:arg_match]{rlang::arg_match} to
improve error messages when function is used internally.}
}
\description{
\itemize{
\item Shift, expand, or contract a bounding box (\code{\link[=sf_bbox_shift]{sf_bbox_shift()}},
\code{\link[=sf_bbox_expand]{sf_bbox_expand()}}, \code{\link[=sf_bbox_contract]{sf_bbox_contract()}})
}
}
