% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_list.R
\name{sf_list}
\alias{sf_list}
\alias{as_sf_list}
\alias{new_sf_list}
\alias{is_sf_list}
\alias{sf_list_rbind}
\alias{map_as_sf}
\alias{map_as_sf_list}
\title{Creating and checking sf lists}
\usage{
as_sf_list(
  x,
  nm = "data",
  col = NULL,
  crs = NULL,
  clean_names = TRUE,
  .name_repair = "check_unique",
  call = caller_env()
)

new_sf_list(
  x,
  nm = "data",
  col = NULL,
  clean_names = TRUE,
  .name_repair = "check_unique",
  call = caller_env()
)

is_sf_list(x, ext = TRUE, allow_null = FALSE)

sf_list_rbind(x, ...)

map_as_sf_list(x, .f, ...)

map_as_sf(x, .f, ...)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{nm}{For \link{as_sf_list}, name(s) for sf list; defaults to "data". If col
is provided, the values of the grouping column are used as names.}

\item{col}{For \link{as_sf_list}, the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}

\item{crs}{A character or numeric reference to a coordinate reference system
supported by \code{\link[sf:st_crs]{sf::st_crs()}} or another  \code{sf}, \code{sfc}, or \code{bbox} object that
is used to provide crs.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{.name_repair}{One of "unique", "universal", or "check_unique". See
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for the meaning of these options.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{ext}{If \code{TRUE}, check if x is a \code{sf}, \code{sfc}, or \code{bbox} class object or
not; defaults to \code{FALSE}. (used by \link{is_sf})}

\item{allow_null}{If \code{TRUE} and x is \code{NULL}, return \code{TRUE}; defaults to \code{FALSE}.}

\item{...}{For \code{\link[=sf_list_rbind]{sf_list_rbind()}}, additional parameters passed to
\code{\link[purrr:list_c]{purrr::list_rbind()}}. For \code{\link[=map_as_sf]{map_as_sf()}}, additional parameters passed to
map.}

\item{.f}{A function, specified in one of the following ways:
\itemize{
\item A named function, e.g. \code{mean}.
\item An anonymous function, e.g. \verb{\\(x) x + 1} or \code{function(x) x + 1}.
\item A formula, e.g. \code{~ .x + 1}. You must use \code{.x} to refer to the first
argument. Only recommended if you require backward compatibility with
older versions of R.
\item A string, integer, or list, e.g. \code{"idx"}, \code{1}, or \code{list("idx", 1)} which
are shorthand for \verb{\\(x) pluck(x, "idx")}, \verb{\\(x) pluck(x, 1)}, and
\verb{\\(x) pluck(x, "idx", 1)} respectively. Optionally supply \code{.default} to
set a default value if the indexed element is \code{NULL} or does not exist.
}}
}
\description{
Functions for coercing an object to an sf_list class objects created by
\code{\link[vctrs:new_list_of]{vctrs::new_list_of()}} and checking lists of sf objects. Any function with
the \code{allow_list} parameter supports this type of input.
}
\seealso{
\code{\link[=is_sf]{is_sf()}}
}
