% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_list.R
\name{sf_list}
\alias{sf_list}
\alias{as_sf_list}
\alias{new_sf_list}
\alias{is_sf_list}
\title{Creating and checking sf lists}
\usage{
as_sf_list(
  x,
  nm = "data",
  col = NULL,
  crs = NULL,
  clean_names = TRUE,
  .name_repair = "check_unique",
  call = caller_env()
)

new_sf_list(
  x,
  nm = "data",
  col = NULL,
  clean_names = TRUE,
  .name_repair = "check_unique",
  call = caller_env()
)

is_sf_list(x, ext = TRUE, allow_null = FALSE)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{nm}{For \link{as_sf_list}, name(s) for sf list; defaults to "data". If col
is provided, the values of the grouping column are used as names.}

\item{col}{For \link{as_sf_list}, the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{ext}{If \code{TRUE}, check if x is a \code{sf}, \code{sfc}, or \code{bbox} class object or
not; defaults to \code{FALSE}. (used by \link{is_sf})}

\item{allow_null}{If \code{TRUE} and x is \code{NULL}, return \code{TRUE}; defaults to \code{FALSE}.}
}
\description{
Functions for coercing an object to an sf_list class objects created by
\code{\link[vctrs:new_list_of]{vctrs::new_list_of()}} and checking lists of sf objects. Any function with
the \code{allow_list} parameter supports this type of input.
}
\seealso{
\code{\link[=is_sf]{is_sf()}}
}
