% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_df.R
\name{sf_to_df}
\alias{sf_to_df}
\alias{df_to_sf}
\alias{check_coords}
\alias{has_coords}
\title{Convert between simple feature and data frame objects}
\usage{
sf_to_df(
  x,
  crs = 4326,
  coords = c("lon", "lat"),
  geometry = "centroid",
  keep_all = TRUE
)

df_to_sf(
  x,
  crs = NULL,
  coords = c("lon", "lat"),
  from_crs = 4326,
  into = NULL,
  sep = ",",
  rev = TRUE,
  remove_coords = FALSE,
  geo = FALSE,
  address = "address"
)

check_coords(x = NULL, coords = NULL, default = c("lon", "lat"), rev = FALSE)

has_coords(x, coords = NULL, value = TRUE)
}
\arguments{
\item{x}{A `sf` or `sfc` object or a data frame with lat/lon coordinates in a
single column or two separated columns.}

\item{crs}{Cordinate reference system to return, Default: 4326 for [sf_to_df]
and NULL for [df_to_sf]}

\item{coords}{Coordinate columns for input dataframe or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{geometry}{Type of geometry to include in data frame. options include
"drop", "wkt", "centroid", "point", Default: 'centroid'.}

\item{keep_all}{If `FALSE`, drop all columns other than those named in
coords, Default: `TRUE`.}

\item{from_crs}{For [df_to_sf], coordinate reference system used by
coordinates or well known text in data frame.}

\item{into}{If coords is a single column name with both longitude and
latitude, `into` is used as the names of the new columns that coords is
separated into. Passed to [tidyr::separate].}

\item{sep}{If coords is a single column name with both longitude and
latitude, `sep` is used as the separator between coordinate values. Passed
to [tidyr::separate].}

\item{rev}{If `TRUE`, reverse c("lat", "lon") coords to c("lon", "lat").
[check_coords] only.}

\item{remove_coords}{For [df_to_sf], if `TRUE`, remove the coordinate columns
after converting a data frame to simple feature object; defaults to
`FALSE`.}

\item{geo}{If `TRUE`, use [address_to_sf] to geocode address column; defaults
to `FALSE`.}

\item{address}{Address column name passed to [tidygeocoder::geocode] or [tidygeocoder::geo]}

\item{default}{c("lon", "lat").}

\item{value}{If TRUE, return the value of the coordinate column names. Used by [has_coords].}
}
\value{
[sf_to_df()] returns a data frame with geometry dropped or converted
  to wkt or coordinates for the centroid or point on surface; [df_to_sf()]
  returns a simple feature object with POINT geometry.
}
\description{
Helper function to convert a simple feature object to data frame by dropping
geometry, converting geometry to well known text, or (if the geometry type is
not POINT) getting coordinates for a centroid or point on surface. If an sfc
object is provided,the "drop" geometry option is not supported.
}
\details{
[check_coords()] is a helper function used by [df_to_sf()] to suggest the
appropriate coordinate column names based on the column names in the provided
data frame.
}
\seealso{
[sf::st_coordinates()]

[ggspatial::df_spatial()]
 [sf::st_as_sf()]
}
