% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfext-package.R
\docType{package}
\name{sfext-package}
\alias{sfext}
\alias{sfext-package}
\title{sfext: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
