% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter_ext.R
\name{st_filter_ext}
\alias{st_filter_ext}
\title{Filter, crop, trim, or erase a simple feature object or list}
\usage{
st_filter_ext(
  x,
  y = NULL,
  crop = FALSE,
  trim = FALSE,
  erase = FALSE,
  crs = NULL,
  .predicate = sf::st_intersects,
  type = NULL,
  list.ok = TRUE,
  ...
)
}
\arguments{
\item{x, y}{A \code{sf}, \code{sfc}, or \code{bbox} object. x may also be a \code{sf} list
objects. If x is an \code{sf} list, additional parameters in \code{...} will be
ignored.}

\item{crop}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[=st_trim]{st_trim()}}.}

\item{erase}{If \code{TRUE}, x is erased by y with \code{\link[=st_erase]{st_erase()}}.}

\item{crs}{Coordinate reference system to return.}

\item{.predicate}{geometry predicate function with the same profile as
\code{\link[sf:geos_binary_pred]{sf::st_intersects()}}; see details for \code{\link[sf:st_join]{sf::st_filter()}} for more options.}

\item{type}{Character string passed to type argument of \code{\link[sf:st_is]{sf::st_is()}}
to filter features to only those matching the specified geometry type.}

\item{list.ok}{If \code{TRUE}, x can be a list of \code{sf}, \code{sfc}, or \code{bbox} objects.
If \code{FALSE}, only \code{sf}, \code{sfc}, or \code{bbox} objects are supported. Defaults to
\code{TRUE}.}

\item{...}{
  Arguments passed on to \code{\link[sf:st_join]{sf::st_filter}}
  \describe{
    \item{\code{}}{}
  }}

\item{null.ok}{If \code{y = NULL} and \code{null.ok = TRUE}, x is returned without
changes. Defaults to \code{TRUE}.}
}
\description{
Extend \code{\link[sf:st_join]{sf::st_filter()}} to filter a sf list or a \code{sf}, \code{sfc}, or \code{bbox} with
options to crop, trim or erase the geometry of the input object based on a
predicate function.
}
