% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_join_ext.R
\name{st_join_ext}
\alias{st_join_ext}
\title{Complete a spatial join using a simple feature objects or an object and list}
\usage{
st_join_ext(x, y, col = NULL, .id = "name", join = NULL, ...)
}
\arguments{
\item{col}{Column name used to convert y into a \code{sf} list if it is not
already.}

\item{.id}{Column name with the values that should be added as a column to
the input \code{sf} object.}

\item{join}{geometry predicate function; defaults to \code{NULL}, set to
\link[sf:geos_binary_pred]{sf::st_intersects} if key_list contains only POLYGON or MULTIPOLYGON
objects or \link[sf:st_nearest_feature]{sf::st_nearest_feature} if key_list contains other types.}

\item{...}{Additional parameters passed to \link[sf:st_join]{sf::st_join}}

\item{boundary}{An sf object with a column named "name" or a list of sf
objects where all items in the list have a "name" column.}
}
\description{
Wrapper for \link[sf:st_join]{sf::st_join} that works with sf lists.
}
