% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_make_grid_ext.R
\name{st_make_grid_ext}
\alias{st_make_grid_ext}
\title{Make a grid over a location bounding box}
\usage{
st_make_grid_ext(
  x,
  crs = NULL,
  cols = NULL,
  rows = NULL,
  gutter = 0,
  desc = FALSE,
  n = NULL,
  cellsize = NULL,
  what = NULL,
  style = "rect",
  .id = "id"
)
}
\arguments{
\item{x}{An object sf, bbox, sfc, raster, or sp object or a data frame that
can be converted to an sf object or a list of sf, bbox, or sfc objects.
st_bbox_asp also supports vectors in the same format as a bbox object.}

\item{crs}{Coordinate reference system of bounding box to return}

\item{cols, rows}{Used to set n if either are not \code{NULL}; defaults to \code{NULL}.
row and id are added as columns to the grid if they are provided.}

\item{gutter}{Distance in units between each column cell; gutter effectively
serves as a margin as the negative buffer is applied to all cells
(including those at the edges of the grid).}

\item{desc}{If TRUE, reverse standard order of cell id numbering; defaults
\code{FALSE}}

\item{n}{If n is NULL and square is \code{TRUE}, the grid is set automatically to
be 10 cells wide, Default: \code{NULL}}

\item{cellsize}{target cellsize}

\item{what}{"polygons", "corners", "centers"; set to centers automatically if
style is "circle", "circle_offset" but a buffer is applied to return
circular polygons.}

\item{style}{Style of cell to return with options including "rect", "square",
"hex", "flat_top_hex", "circle", "circle_offset"}

\item{.id}{A name to use for the cell id column.}

\item{location}{A \code{sf}, \code{sfc}, or \code{bbox} object, Default: \code{NULL}. Required.}
}
\description{
Create a grid with an id column and optionally a set number of columns and
rows. This documentation is incomplete the function may change.
}
\examples{

}
\seealso{
\link[sf:st_make_grid]{sf::st_make_grid}
}
