% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_make_grid_ext.R
\name{st_make_grid_ext}
\alias{st_make_grid_ext}
\title{Make a grid over a simple feature bounding box}
\usage{
st_make_grid_ext(
  x,
  ...,
  unit = NULL,
  crs = NULL,
  ncol = NULL,
  nrow = NULL,
  n = NULL,
  gutter = 0,
  desc = FALSE,
  cellsize = NULL,
  what = NULL,
  style = "rect",
  .id = "id",
  filter = FALSE,
  trim = FALSE
)
}
\arguments{
\item{x}{A \code{sf}, \code{sfc}, or \code{bbox} object, Default: \code{NULL}. Required.}

\item{...}{
  Arguments passed on to \code{\link[=st_bbox_ext]{st_bbox_ext}}
  \describe{
    \item{\code{nudge}}{Passed as to parameter \code{\link[=st_nudge]{st_nudge()}} when not \code{NULL}. Numeric
vector or sf object.}
    \item{\code{dist}}{buffer distance in units. Optional.}
    \item{\code{diag_ratio}}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}
    \item{\code{allow_list}}{If \code{TRUE}, allow sf list objects as an input and use
\code{\link[purrr:map]{purrr::map()}} to apply the provided parameters to each object within the
list to return as a new sf list object.}
    \item{\code{asp}}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[=get_asp]{get_asp()}} returns the same value
without modification.}
  }}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{crs}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}

\item{ncol, nrow}{Used to set n if either are not \code{NULL}; defaults to \code{NULL}.
row and id are added as columns to the grid if they are provided.}

\item{n}{If n is NULL and square is \code{TRUE}, the grid is set automatically to
be 10 cells wide, Default: \code{NULL}}

\item{gutter}{Distance in units between each column cell; gutter effectively
serves as a margin as the negative buffer is applied to all cells
(including those at the edges of the grid).}

\item{desc}{If TRUE, reverse standard order of cell id numbering; defaults
\code{FALSE}}

\item{cellsize}{target cellsize}

\item{what}{"polygons", "corners", "centers"; set to centers automatically if
style is "circle", "circle_offset" but a buffer is applied to return
circular polygons.}

\item{style}{Style of cell to return with options including "rect", "square",
"hex", "flat_top_hex", "circle", "circle_offset"}

\item{.id}{A name to use for the cell id column. Defaults to "id".}

\item{filter}{If \code{TRUE} (or if trim is \code{TRUE}) filter grid geometry by x
using \link{st_filter_ext}}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[=st_trim]{st_trim()}}.}
}
\description{
Create a grid with an id column and optionally a set number of columns and
rows. This documentation is incomplete the function may change.
}
\examples{
nc <- read_sf_ext(system.file("shape/nc.shp", package = "sf"))

# Make a 2 by 2 grid across a location with a 1000 meter gutter between each cell
plot(
  st_make_grid_ext(
    x = nc[24, ],
    dist = 500,
    unit = "meter",
    ncol = 2,
    nrow = 2,
    gutter = 1000
  )
)

# Make a 5 by 5 grid with a 8.5 by 11 aspect ratio filtered to x
plot(
  st_make_grid_ext(
    x = nc[24, ],
    asp = 8.5 / 11,
    ncol = 5,
    nrow = 5,
    filter = TRUE
  )
)

# Make a 4 by 5 grid of circles trimmed to x boundaries
plot(
  st_make_grid_ext(
    x = nc[24, ],
    ncol = 4,
    nrow = 5,
    style = "circle_offset",
    trim = TRUE
  ),
  max.plot = 1
)
}
\seealso{
\link[sf:st_make_grid]{sf::st_make_grid}
}
