% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nudge.R
\name{st_nudge}
\alias{st_nudge}
\title{Nudge a simple feature to the center of another feature and/or a set distance}
\usage{
st_nudge(
  x,
  to = NULL,
  nudge_y = 0,
  nudge_x = 0,
  unit = NULL,
  scale = 1,
  rotate = 0,
  crs = NULL
)
}
\arguments{
\item{x}{A \code{sf}, \code{sfc}, or \code{bbox} object}

\item{to}{sf object to use as new center for x or length 2 numeric vector
with the nudge_y and nudge_x distance (in that order).}

\item{nudge_y, nudge_x}{Distance to nudge geometry in unit. If unit is \code{NULL},
distance is assumed to be in the same units as the coordinate reference
system of the input object.}

\item{unit}{Units for nudge_y and nudge_x distance (also used if to is
numeric).}

\item{scale}{numeric; scale factor, Default: 1}

\item{rotate}{numeric; degrees to rotate (-360 to 360), Default: 0}
}
\description{
Nudge, move, or shift a \code{sf}, \code{sfc}, or \code{bbox} object to the center of
another feature and/or by a set distance.
}
