% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nudge.R
\name{st_nudge}
\alias{st_nudge}
\alias{st_nudge.default}
\alias{st_nudge.bbox}
\alias{st_nudge.sf}
\title{Nudge a simple feature to the center of another feature and/or a set distance}
\usage{
st_nudge(
  x,
  to = NULL,
  nudge_y = 0,
  nudge_x = 0,
  unit = NULL,
  scale = 1,
  rotate = 0,
  crs = NULL
)

\method{st_nudge}{default}(
  x,
  to = NULL,
  nudge_y = 0,
  nudge_x = 0,
  unit = NULL,
  scale = 1,
  rotate = 0,
  crs = NULL
)

\method{st_nudge}{bbox}(x, ...)

\method{st_nudge}{sf}(x, ...)
}
\arguments{
\item{x}{A \code{sf}, \code{bbox}, \code{sfc}, \code{raster}, \code{sp}, or data frame object that can
be converted into a simple feature or bounding box object. \code{\link[=as_bbox]{as_bbox()}} can
also convert a vector with xmin, ymin, xmax, and ymax values.
\code{\link[=as_sf_list]{as_sf_list()}} only supports sf objects or a data frames with a sf list
column named "data" (typically created by using \code{\link[dplyr:group_nest]{dplyr::group_nest()}} on an
sf object.}

\item{to}{sf object to use as new center for x or length 2 numeric vector
with the nudge_y and nudge_x distance (in that order).}

\item{nudge_y, nudge_x}{Distance to nudge geometry in unit. If unit is \code{NULL},
distance is assumed to be in the same units as the coordinate reference
system of the input object.}

\item{unit}{Units for nudge_y and nudge_x distance (also used if to is
numeric).}

\item{scale}{numeric; scale factor, Default: 1}

\item{rotate}{numeric; degrees to rotate (-360 to 360), Default: 0}

\item{crs}{Coordinate reference system for \code{sf}, \code{bbox}, \code{sfc} or \code{sf} list
object to return.}

\item{...}{Additional parameters passed to \link{as_sf}, \link{as_sfc}, \link{as_bbox}, or
\link{as_sf_list}}
}
\description{
Nudge, move, or shift a \code{sf}, \code{sfc}, or \code{bbox} object to the center of
another feature and/or by a set distance.
}
