% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_transform_ext.R
\name{st_transform_ext}
\alias{st_transform_ext}
\alias{st_transform_omerc}
\alias{st_omerc}
\title{Transform or convert coordinates of a simple feature or bounding box object}
\usage{
st_transform_ext(
  x,
  crs = NULL,
  class = NULL,
  rotate = 0,
  null.ok = FALSE,
  list.ok = TRUE
)

st_omerc(x, rotate = 0)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object, a list of sf objects.}

\item{crs}{A character or numeric reference to a coordinate reference system
supported by \code{\link[sf:st_crs]{sf::st_crs()}} or another  \code{sf}, \code{sfc}, or \code{bbox} object that
is used to provide crs.}

\item{class}{Class of object to return (\code{sf} or \code{bbox}). If x is an sf list,
the returned object remains a list but may be converted to \code{bbox} if class
= "sf".}

\item{rotate}{If rotate is greater or less than 0, \link{st_transform_ext} calls
\link{st_omerc} and returns an object with the oblique mercator projection
passing the value of rotate to the gamma parameter of the projection.
rotate must be between -45 and 45 degrees.}
}
\value{
An \code{sf}, \code{sfc}, or \code{bbox} object transformed to a new coordinate
reference system.
}
\description{
Take a \code{sf}, \code{sfc}, or \code{bbox} object and transform to coordinate reference
system to match the object provided to \code{crs}.
}
\seealso{
\code{\link[sf:st_transform]{sf::st_transform()}},\code{\link[sf:st_crs]{sf::st_crs()}}
}
