% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_transform_ext.R
\name{st_transform_ext}
\alias{st_transform_ext}
\alias{st_omerc}
\alias{st_transform_omerc}
\alias{st_wgs84}
\alias{as_wgs84}
\title{Transform or convert coordinates of a simple feature or bounding box object}
\usage{
st_transform_ext(
  x,
  crs = NULL,
  class = NULL,
  rotate = 0,
  allow_null = FALSE,
  allow_list = TRUE
)

st_omerc(x, rotate = 0)

st_wgs84(x)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{bbox} object, a list of sf objects, or a
\code{data.frame} object (always returned as is).}

\item{crs}{A character or numeric reference to a coordinate reference system
supported by \code{\link[sf:st_crs]{sf::st_crs()}} or another  \code{sf}, \code{sfc}, or \code{bbox} object that
is used to provide crs.}

\item{class}{Class of object to return (\code{sf} or \code{bbox}). If x is an sf list,
the returned object remains a list but may be converted to \code{bbox} if \code{class = "sf"}.}

\item{rotate}{If rotate is greater or less than 0, \code{\link[=st_transform_ext]{st_transform_ext()}} calls
\code{\link[=st_omerc]{st_omerc()}} and returns an object with the Oblique Mercator projection
passing the value of rotate to the gamma parameter of the projection.
rotate must be between -45 and 45 degrees.}

\item{allow_null}{If \code{TRUE} and x is \code{NULL} return x without an error.
Defaults to \code{FALSE}.}

\item{allow_list}{If \code{TRUE}, x can be a list of \code{sf}, \code{sfc}, or \code{bbox} objects.
If \code{FALSE}, only \code{sf}, \code{sfc}, or \code{bbox} objects are supported. Defaults to
\code{TRUE}.}
}
\value{
An \code{sf}, \code{sfc}, or \code{bbox} object transformed to a new coordinate
reference system.
}
\description{
This function wraps \code{\link[sf:st_transform]{sf::st_transform()}} but supports a wider range of input
objects and, using the \code{\link[=as_sf_class]{as_sf_class()}} function, returns a wider range of
objects. Typically, takes a \code{sf}, \code{sfc}, or \code{bbox} object and transform to
coordinate reference system to match the value of crs or the object provided
to crs. If \code{x} is a data.frame or if \code{x} is \code{NULL} and \code{allow_null} is \code{TRUE}
(defaults to \code{FALSE}) it is returned as is.
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))

nc_bbox <- sf::st_bbox(nc)

nc_3857 <- st_transform_ext(nc, 3857)

st_transform_ext(nc_bbox, crs = 4326)

sf::st_crs(st_transform_ext(nc_3857, crs = nc))$input

sf::st_crs(st_wgs84(nc_3857))$input

}
\seealso{
\code{\link[sf:st_transform]{sf::st_transform()}},\code{\link[sf:st_crs]{sf::st_crs()}}
}
