% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_union_ext.R
\name{st_union_ext}
\alias{st_union_ext}
\title{Union simple feature objects and combine name column values}
\usage{
st_union_ext(
  x = NULL,
  y = NULL,
  name_col = "name",
  sf_col = "geometry",
  label = NULL,
  ext = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{sf}, \code{sfc}, or \code{bbox} object, Default: \code{NULL}}

\item{y}{A \code{sf} or \code{sfc} object, Default: \code{NULL}}

\item{name_col}{Column name to collapse into new name_col value, Default:
'name'}

\item{sf_col}{Geometry column name to use if x is an sf object, Default:
'geometry'}

\item{label}{Length 1 character vector used if name_col is \code{NULL}.}

\item{ext}{If \code{FALSE}, st_union_ext functions the same as \code{\link[sf:geos_combine]{sf::st_union()}},
Default: \code{TRUE}}

\item{...}{Additional parameters passed to \code{\link[sf:geos_combine]{sf::st_union()}}}
}
\value{
A sfc object if y is \code{NULL} and ext is \code{FALSE}. A tibble sf if x is a
sf object and y is \code{NULL}. If y is provided, \code{\link[=st_union_ext]{st_union_ext()}} is identical to
\code{\link[sf:geos_combine]{sf::st_union()}}
}
\description{
Wrapper for \code{\link[sf:geos_combine]{sf::st_union()}} supporting the additional feature of a combined
name column collapsed into a single vector with \code{\link[cli:pluralize]{cli::pluralize()}}.
}
\examples{
nc <- read_sf_path(system.file("shape/nc.shp", package = "sf"))
nc_union <- st_union_ext(nc[10:15,], name_col = "NAME")
nc_union
plot(nc_union)
}
