% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_union_ext.R
\name{st_union_ext}
\alias{st_union_ext}
\title{Union simple feature objects}
\usage{
st_union_ext(
  x = NULL,
  y = NULL,
  name_col = "name",
  sf_col = "geometry",
  ext = TRUE,
  ...
)
}
\arguments{
\item{x}{A sf, sfc, or bbox object, Default: \code{NULL}}

\item{y}{A sf or sfc object, Default: \code{NULL}}

\item{name_col}{Column name to collapse into new name_col value, Default:
'name'}

\item{sf_col}{Geometry column name to use if x is an sf object, Default:
'geometry'}

\item{ext}{If \code{FALSE}, st_union_ext functions the same as \link[sf:geos_combine]{sf::st_union},
Default: \code{TRUE}}

\item{...}{Additional parameters passed to \link[sf:geos_combine]{sf::st_union}}
}
\value{
A sfc object if y is \code{NULL} and ext is \code{FALSE}. A tibble sf if x is a
sf object and y is \code{NULL}. If y is provided, \link{st_union_ext} is identical to
\link[sf:geos_combine]{sf::st_union}
}
\description{
Wrapper for \link[sf:geos_combine]{sf::st_union} supporting the additional feature of a combined
name column collapsed into a single vector with \link[cli:pluralize]{cli::pluralize}.
}
