% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_exif.R
\name{write_exif_from}
\alias{write_exif_from}
\alias{write_exif_keywords}
\title{Write EXIF data for photos on spatial join with a sf object or list of sf objects}
\usage{
write_exif_from(
  path,
  fileext = NULL,
  filetype = NULL,
  from,
  .id = "name",
  tag = "keywords",
  join = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{path}{A path to folder or file.}

\item{fileext, filetype}{File extension or file type. filetype is used if
fileext is \code{NULL}.}

\item{from}{A sf object or list of sf objects where each object has a column
with a name matching the .id parameter. The attribute value in this column
are used to assign the tag parameter to the file at the provided path based
on the spatial relationship set by join. For example, from may be boundary
data used to assign keywords based on photo locations.}

\item{.id}{Column name in from with the values to use for tag values.}

\item{tag}{EXIF tag to update, supported options include "keywords", "title",
or "description".}

\item{join}{geometry predicate function; defaults to \code{NULL}, set to
\link[sf:geos_binary_pred]{sf::st_intersects} if from contains only POLYGON or MULTIPOLYGON objects
or \link[sf:st_nearest_feature]{sf::st_nearest_feature} if from contains other types.}

\item{overwrite}{If \code{TRUE}, overwrite any existing EXIF metadata present in
the provided fields; defaults to \code{TRUE}}
}
\description{
Extends \code{\link[=read_sf_exif]{read_sf_exif()}} and \code{\link[filenamr:read_exif]{filenamr::write_exif()}}
}
